% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-datasets.R
\name{addBinomialRate}
\alias{addBinomialRate}
\title{Calculate a simulated set of expected proportions based on imports, growth rates and a baseline total}
\usage{
addBinomialRate(growthSimulation, baselineExpr = rpois(n(), Est.actual))
}
\arguments{
\item{growthSimulation}{- a timeseries with an Growth.actual column and a import column.}

\item{baselineExpr}{- an expression to derive the baseline total - either a single number or a way of calculating that number from growthSimulation}
}
\value{
the timeseries with an expected binomial proportion in the column Proportion.actual.
}
\description{
Calculate a simulated set of expected proportions based on imports, growth rates and a baseline total
}
