% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-datasets.R
\name{addBootstrappedObservations}
\alias{addBootstrappedObservations}
\title{Generate a random set of observations from a poisson rate}
\usage{
addBootstrappedObservations(
  growthSimulation,
  bootstraps = 100,
  delayMean = 0,
  delaySd = 1,
  lastObservation = nrow(growthSimulation$ts) - 1
)
}
\arguments{
\item{growthSimulation}{- the growth_simulation}

\item{bootstraps}{- the number of generated bootstraps}

\item{delayMean}{}

\item{delaySd}{}

\item{lastObservation}{}
}
\value{
a growth_simulation with a time series including bootstrapped observations (as "value" column, bootstrap id in "subgroup")
}
\description{
Generate a random set of observations from a poisson rate
}
