% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-datasets.R
\name{addImportations}
\alias{addImportations}
\title{Add importations to the simulation. Without importations there will be no growth so some cases must be imported.}
\usage{
addImportations(
  growthSimulation,
  importDf = tibble(time = 0, import = rate),
  rate = 100
)
}
\arguments{
\item{growthSimulation}{- a timeseries with an integer time column.}

\item{importDf}{- an optional dataframe with 2 columns: "time" and "import" counts.}

\item{rate}{- if no importDf is specified the rate parameter will define the number of cases on the zero day.}
}
\value{
a growth_simulation with an additional import column
}
\description{
Add importations to the simulation. Without importations there will be no growth so some cases must be imported.
}
