% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-datasets.R
\name{addObservedRate}
\alias{addObservedRate}
\title{Add in a simulated rate of observed cases based on the underlying growth rate and imports with a weekly variation}
\usage{
addObservedRate(
  growthSimulation,
  observedFractionExpr = 1,
  weekendEffect = 0.1
)
}
\arguments{
\item{growthSimulation}{- the growth rate simulation}

\item{observedFractionExpr}{- what proportion of the cases are observed, either as a number, a vector of same length as the simulation, or a way of calculating that number from the growthSimulation. Typically under 1.}

\item{weekendEffect}{- what fraction of cases can we expect to not be observed on a saturday or sunday.}
}
\value{
the timeseries with an expected observed rate in the column Est.observed
}
\description{
Add in a simulated rate of observed cases based on the underlying growth rate and imports with a weekly variation
}
