% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-datasets.R
\name{addPoissonRate}
\alias{addPoissonRate}
\title{Calculate a simulated set of expected incidences based on imports and growth rate}
\usage{
addPoissonRate(growthSimulation)
}
\arguments{
\item{growthSimulation}{- a timeseries with an Growth.actual column and a import column.}
}
\value{
the timeseries with an expected poisson rate column Est.actual.
}
\description{
Calculate a simulated set of expected incidences based on imports and growth rate
}
