% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{AnotherTestRApi}

\alias{AnotherTestRApi}

\title{}

\usage{\preformatted{
library(testRapi)
J = JavaApi$new();
instance <- J$AnotherTestRApi$new(...);
}}

\description{
	This has no documentation
}

\details{
	This has no documentation
}

\examples{
## -----------------------------------
## Method `AnotherTestRApi$create`
## -----------------------------------

## -----------------------------------
## Method `AnotherTestRApi$concat`
## -----------------------------------

## -----------------------------------
## Method `AnotherTestRApi$throwCatchable`
## -----------------------------------

## -----------------------------------
## Method `AnotherTestRApi$printMessage`
## -----------------------------------

## -----------------------------------
## Method `AnotherTestRApi$throwRuntime`
## -----------------------------------

}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$AnotherTestRApi$new(message1, message2)}}
		}
	}
	\subsection{Static methods}{
		\itemize{
			\item \href{#method-create}{\code{J$AnotherTestRApi$create(message1, message2)}}
			\item \href{#method-concat}{\code{J$AnotherTestRApi$concat(message1, message2)}}
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-throwCatchable}{\code{instance$throwCatchable()}}
			\item \href{#method-printMessage}{\code{instance$printMessage()}}
			\item \href{#method-throwRuntime}{\code{instance$throwRuntime()}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		the first constructor is used if there are none annotated
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$AnotherTestRApi$new(message1, message2)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message1 - the message to be printed}{ - (java expects a java.lang.String)}
				\item{message2 - will be used for toString}{ - (java expects a java.lang.String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 AnotherTestRApi object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-create"></a>}}
	
	\subsection{Method \code{create()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$AnotherTestRApi$create(message1, message2)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message1}{ - (java expects a java.lang.String)}
				\item{message2}{ - (java expects a java.lang.String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 AnotherTestRApi object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-concat"></a>}}
	
	\subsection{Method \code{concat()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$AnotherTestRApi$concat(message1, message2)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message1}{ - (java expects a java.lang.String)}
				\item{message2}{ - (java expects a java.lang.String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.String: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-throwCatchable"></a>}}
	
	\subsection{Method \code{throwCatchable()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$AnotherTestRApi$new(...);
  instance$throwCatchable()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.String: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-printMessage"></a>}}
	
	\subsection{Method \code{printMessage()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$AnotherTestRApi$new(...);
  instance$printMessage()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-throwRuntime"></a>}}
	
	\subsection{Method \code{throwRuntime()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$AnotherTestRApi$new(...);
  instance$throwRuntime()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.String: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}