% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{BounceTest}

\alias{BounceTest}

\title{BounceTest}

\usage{\preformatted{
library(testRapi)
J = JavaApi$new();
instance <- J$BounceTest$new(...);
}}

\description{
	
}

\details{
	
}

\examples{
## -----------------------------------
## Method `BounceTest$bounceVoid`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceInt`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceDouble`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceString`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceNull`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceCharacter`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceCharacterVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceInteger`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceIntegerVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceDate`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceDateVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceFactor`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceFactorVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceLogical`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceLogicalVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceNumeric`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceNumericVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceDataframe`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceList`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceNamedList`
## -----------------------------------

}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$BounceTest$new()}}
		}
	}
	\subsection{Static methods}{
		\itemize{
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-bounceVoid}{\code{instance$bounceVoid()}}
			\item \href{#method-bounceInt}{\code{instance$bounceInt(x)}}
			\item \href{#method-bounceDouble}{\code{instance$bounceDouble(x)}}
			\item \href{#method-bounceString}{\code{instance$bounceString(x)}}
			\item \href{#method-bounceNull}{\code{instance$bounceNull(x)}}
			\item \href{#method-bounceCharacter}{\code{instance$bounceCharacter(x)}}
			\item \href{#method-bounceCharacterVector}{\code{instance$bounceCharacterVector(x)}}
			\item \href{#method-bounceInteger}{\code{instance$bounceInteger(x)}}
			\item \href{#method-bounceIntegerVector}{\code{instance$bounceIntegerVector(x)}}
			\item \href{#method-bounceDate}{\code{instance$bounceDate(x)}}
			\item \href{#method-bounceDateVector}{\code{instance$bounceDateVector(x)}}
			\item \href{#method-bounceFactor}{\code{instance$bounceFactor(x)}}
			\item \href{#method-bounceFactorVector}{\code{instance$bounceFactorVector(x)}}
			\item \href{#method-bounceLogical}{\code{instance$bounceLogical(x)}}
			\item \href{#method-bounceLogicalVector}{\code{instance$bounceLogicalVector(x)}}
			\item \href{#method-bounceNumeric}{\code{instance$bounceNumeric(x)}}
			\item \href{#method-bounceNumericVector}{\code{instance$bounceNumericVector(x)}}
			\item \href{#method-bounceDataframe}{\code{instance$bounceDataframe(x)}}
			\item \href{#method-bounceList}{\code{instance$bounceList(x)}}
			\item \href{#method-bounceNamedList}{\code{instance$bounceNamedList(x)}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$BounceTest$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 BounceTest object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceVoid"></a>}}
	
	\subsection{Method \code{bounceVoid()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceVoid()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@7672960e: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceInt"></a>}}
	
	\subsection{Method \code{bounceInt()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceInt(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@163df259)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@163df259: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDouble"></a>}}
	
	\subsection{Method \code{bounceDouble()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceDouble(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@56b751b1)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@56b751b1: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceString"></a>}}
	
	\subsection{Method \code{bounceString()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceString(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@5a78b52b)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@5a78b52b: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNull"></a>}}
	
	\subsection{Method \code{bounceNull()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceNull(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@2f61de04)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@2f61de04: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceCharacter"></a>}}
	
	\subsection{Method \code{bounceCharacter()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceCharacter(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@40013051)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@40013051: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceCharacterVector"></a>}}
	
	\subsection{Method \code{bounceCharacterVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceCharacterVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@5f5a33ed)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@5f5a33ed: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceInteger"></a>}}
	
	\subsection{Method \code{bounceInteger()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceInteger(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@5b59c3d)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@5b59c3d: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceIntegerVector"></a>}}
	
	\subsection{Method \code{bounceIntegerVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceIntegerVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@69e36fe3)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@69e36fe3: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDate"></a>}}
	
	\subsection{Method \code{bounceDate()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceDate(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@48d31d25)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@48d31d25: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDateVector"></a>}}
	
	\subsection{Method \code{bounceDateVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceDateVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@46188a89)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@46188a89: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceFactor"></a>}}
	
	\subsection{Method \code{bounceFactor()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceFactor(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@6f65a203)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@6f65a203: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceFactorVector"></a>}}
	
	\subsection{Method \code{bounceFactorVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceFactorVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@1d0b447b)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@1d0b447b: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceLogical"></a>}}
	
	\subsection{Method \code{bounceLogical()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceLogical(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@4de158ca)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@4de158ca: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceLogicalVector"></a>}}
	
	\subsection{Method \code{bounceLogicalVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceLogicalVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@4f2fdf3d)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@4f2fdf3d: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNumeric"></a>}}
	
	\subsection{Method \code{bounceNumeric()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceNumeric(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@d180961)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@d180961: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNumericVector"></a>}}
	
	\subsection{Method \code{bounceNumericVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceNumericVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@4af7ac25)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@4af7ac25: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDataframe"></a>}}
	
	\subsection{Method \code{bounceDataframe()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceDataframe(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@d978ab9)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@d978ab9: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceList"></a>}}
	
	\subsection{Method \code{bounceList()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceList(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@3cf06a01)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@3cf06a01: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNamedList"></a>}}
	
	\subsection{Method \code{bounceNamedList()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceNamedList(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@71ee2fff)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@71ee2fff: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}