% Generated by r6-generator-maven-plugin: do not edit by hand

\name{CoriEstimator}

\alias{CoriEstimator}

\title{CoriEstimator}

\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{r0Mean}{r0Mean - The mean of the gamma distribution of the baseline prior assumption about R0 - the theorerical reproduction number in an infinite pool of evenly mixing susceptible individuals - (java expects a double)}
	\item{r0SD}{r0SD - The SD of the prior gamma distribution - (java expects a double)}
	\item{maxWindow}{maxWindow - The maximum sliding window over which to estimate \(R_t\). - (java expects a int)}
	\if{html}{\out{</div>}}
}


\description{
An estimator of the reproduction number (\(R_t\)) using a renewal equation method. 

Version: 0.03

Generated: 2022-05-24T22:12:30.975629
}

\details{
	
The renewal equation method depends on a time series of infections, and on the infectivity profile - 
a measure of the probability that a secondary infection occurred on a specific day after the primary case, 
given a secondary infection occurred. A Bayesian framework is then used to update a prior probabilistic 
estimate of \(R_t\) on any given day with both information gained from the time series of infections in the 
epidemic to date and the infectivity profile to produce a posterior estimate of \(R_t\). This is based on the 
article:<br>

<br>A. Cori, N. M. Ferguson, C. Fraser, and S. Cauchemez, ‘A New Framework and Software to Estimate Time-Varying Reproduction Numbers During Epidemics’, Am J Epidemiol, vol. 178, no. 9, pp. 1505–1512, Nov. 2013, doi: 10.1093/aje/kwt133.<br>

<br> and the implementation of it in the R Library EpiEstim. It assumes that infections (or other type of observation) are a 
Poisson distributed process and that the reproduction number is a gamma distributed quantity. The estimates of \(R_t\) may be 
aggregated over a specific "window" to provide a more stable estimate.
}

\examples{\dontrun{
J = jepidemic::JavaApi$get();
instance = J$CoriEstimator$new(r0Mean, r0SD, maxWindow)

## -----------------------------------
## Method `CoriEstimator$legacySupport`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$defaultEpiEstim`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$defaultUncertainEpiEstim`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$selectSpecificWindow`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$selectAdaptiveWindow`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$selectMinimumUncertainty`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$selectMixtureCombination`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$selectWeightedMixtureCombination`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$withDefaultPrior`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$withFixedPrior`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$withAdaptivePrior`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$collectResampledQuantiles`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$collectMixtureQuantiles`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$collectMixtureApproximation`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$detailedOutput`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$withInfectivityProfile`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$withInfectivityProfileMatrix`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$atStartOfTimeseries`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$inMiddleOfTimeseries`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$estimateRt`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$estimateRtSingle`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$estimateRtFromRates`
## -----------------------------------

## -----------------------------------
## Method `CoriEstimator$estimateRtFromSingleRate`
## -----------------------------------

}}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$CoriEstimator$new(r0Mean, r0SD, maxWindow)}}
		}
	}
	\subsection{Static methods}{
		\itemize{
			\item \href{#method-defaultEpiEstim}{\code{J$CoriEstimator$defaultEpiEstim(infectivityProfile, meanR0, sdR0, window)}}
			\item \href{#method-defaultUncertainEpiEstim}{\code{J$CoriEstimator$defaultUncertainEpiEstim(infectivityProfiles, meanR0, sdR0, window, n2)}}
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-legacySupport}{\code{instance$legacySupport(on)}}
			\item \href{#method-selectSpecificWindow}{\code{instance$selectSpecificWindow(window)}}
			\item \href{#method-selectAdaptiveWindow}{\code{instance$selectAdaptiveWindow(incidenceSum, minWindow)}}
			\item \href{#method-selectMinimumUncertainty}{\code{instance$selectMinimumUncertainty(timeVsRt, minWindow)}}
			\item \href{#method-selectMixtureCombination}{\code{instance$selectMixtureCombination()}}
			\item \href{#method-selectWeightedMixtureCombination}{\code{instance$selectWeightedMixtureCombination(weights)}}
			\item \href{#method-withDefaultPrior}{\code{instance$withDefaultPrior()}}
			\item \href{#method-withFixedPrior}{\code{instance$withFixedPrior(mean, sd)}}
			\item \href{#method-withAdaptivePrior}{\code{instance$withAdaptivePrior(factor)}}
			\item \href{#method-collectResampledQuantiles}{\code{instance$collectResampledQuantiles(sampleSize)}}
			\item \href{#method-collectMixtureQuantiles}{\code{instance$collectMixtureQuantiles()}}
			\item \href{#method-collectMixtureApproximation}{\code{instance$collectMixtureApproximation()}}
			\item \href{#method-detailedOutput}{\code{instance$detailedOutput()}}
			\item \href{#method-withInfectivityProfile}{\code{instance$withInfectivityProfile(infectivityProfile, replace)}}
			\item \href{#method-withInfectivityProfileMatrix}{\code{instance$withInfectivityProfileMatrix(infectivityProfiles)}}
			\item \href{#method-atStartOfTimeseries}{\code{instance$atStartOfTimeseries()}}
			\item \href{#method-inMiddleOfTimeseries}{\code{instance$inMiddleOfTimeseries()}}
			\item \href{#method-estimateRt}{\code{instance$estimateRt(incidence, dateColName, incidenceColName)}}
			\item \href{#method-estimateRtSingle}{\code{instance$estimateRtSingle(incidence, dateColName, incidenceColName)}}
			\item \href{#method-estimateRtFromRates}{\code{instance$estimateRtFromRates(rates, dateColName, rateColName, samplesPerProfile)}}
			\item \href{#method-estimateRtFromSingleRate}{\code{instance$estimateRtFromSingleRate(rate, dateColName, rateColName, samplesPerProfile)}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		The Estimator is configured with the baseline \(R_0\) assumptions which are used as prior values for \(R_t\) when no other 
values are available, or the {@link CoriEstimator#withDefaultPrior()} option is used.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
J$CoriEstimator$new(r0Mean, r0SD, maxWindow)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{r0Mean - The mean of the gamma distribution of the baseline prior assumption about R0 - the theorerical reproduction number in an infinite pool of evenly mixing susceptible individuals}{ - (java expects a double)}
				\item{r0SD - The SD of the prior gamma distribution}{ - (java expects a double)}
				\item{maxWindow - The maximum sliding window over which to estimate \(R_t\).}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-legacySupport"></a>}}
	
	\subsection{Method \code{legacySupport()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$legacySupport(on)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{on - a boolean flag to switch on (TRUE) or off (FALSE) eipestim naming convention}{ - (java expects a boolean)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: the estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-defaultEpiEstim"></a>}}
	
	\subsection{Method \code{defaultEpiEstim()}}{
		Produces a CoriEstimator that is configured to behave like the default settings of EpiEstim. 
In that it uses a fixed prior for Rt and returns a fixed window size<br><br>
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
J$CoriEstimator$defaultEpiEstim(infectivityProfile, meanR0, sdR0, window)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{infectivityProfile - a vector with the discrete probability distribution of the serial interval with \(P_{(t=0)} = 0\)}{ - (java expects a RNumericVector)}
				\item{meanR0 - the mean prior of R0}{ - (java expects a double)}
				\item{sdR0 - the sd prior of R0}{ - (java expects a double)}
				\item{window - the window size}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: a default cori estimator
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-defaultUncertainEpiEstim"></a>}}
	
	\subsection{Method \code{defaultUncertainEpiEstim()}}{
		Produces a CoriEstimator that is configured to behave like the default settings of EpiEstim. 
In that it uses a fixed prior for Rt and returns a fixed window size and aggregates uncertain SI 
distributions with a random resampling strategy<br><br>
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
J$CoriEstimator$defaultUncertainEpiEstim(infectivityProfiles, meanR0, sdR0, window, n2)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{infectivityProfiles - a numeric matrix with the discrete probability distribution of the serial interval with \(P_{(t=0)} = 0\)}{ - (java expects a RNumericArray)}
				\item{meanR0 - the mean prior of R0}{ - (java expects a double)}
				\item{sdR0 - the sd prior of R0}{ - (java expects a double)}
				\item{window - the window size}{ - (java expects a int)}
				\item{n2 - the number of samples to use for estimating confidence limits on bootstrapped infectivity profiles}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: a default cori estimator
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-selectSpecificWindow"></a>}}
	
	\subsection{Method \code{selectSpecificWindow()}}{
		The select methods define how the posterior estimate of \(R_t\) is selected. In this case a set value for the window is used. This must be less or equal to the 
value of maxWindow supplied in the CoriEstimator::new constructor. The window defines how many data points prior to the time of the estimate are used to synthesise the posterior
estimate of \(R_t\), and so tends to reduce the uncertainty around the value of \(R_t\), but longer values of window introduces temporal uncertainty into
the estimates. This is as implemented in EpiEstim (with a default value of 7 for the window).
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$selectSpecificWindow(window)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{window - a number of days}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-selectAdaptiveWindow"></a>}}
	
	\subsection{Method \code{selectAdaptiveWindow()}}{
		The adaptive windowing strategy selects a posterior estimate of \(R_t\) by selecting the shortest window length (&gt; minWindow) which spans at least incidenceSum infection 
(or observation) counts. This ensures that where numbers are low the estimation window is lengthened (up ot the value of maxWindow provided to the CoriEstimator::new constructor) to 
include more data at the cost of temporal precision. This prevents stochastic noise from dominating the estimates of \(R_t\).
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$selectAdaptiveWindow(incidenceSum, minWindow)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{incidenceSum - the number of cases that must be observed in a window. It there are fewer that this the window will be lengthened. Larger values lead to smoother estiamtes at the cost of reduced temporal precision.}{ - (java expects a double)}
				\item{minWindow - the minimum size of the window that will be selected. In times of very high case numbers a minimum window prevents spurious outliers from completely dominating that days estimate of \(R_t\).}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-selectMinimumUncertainty"></a>}}
	
	\subsection{Method \code{selectMinimumUncertainty()}}{
		The minimum uncertainty strategy aims to select a window for the posterior \(R_t\) estimate based on minimising a combined value for uncertainty. The uncertainty in this case is the 
product of the (timeVsRt)th root of SD of the posterior estimate and the window size to the power of timeVsRt. <br><br>

$$window^{timeVsRt} * SD_{posterior}^\frac{1}{timeVsRt}$$
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$selectMinimumUncertainty(timeVsRt, minWindow)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{timeVsRt - the relative importance parameter. Values above one favour temporal certainty; values under one favour \(R_t\) estimate certainty}{ - (java expects a double)}
				\item{minWindow - the minimum size of the window that will be selected.}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-selectMixtureCombination"></a>}}
	
	\subsection{Method \code{selectMixtureCombination()}}{
		The mixture combination strategy for selecting a posterior \(R_t\) involves finding all the estimates that have a window that span the estimate date (this may involve longer windowed estimate from
dates in the future) and estimating a gamma distribution with the same mean and SD as the mixture of all the estimates. This combined estimate involves estimates of all different windows.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$selectMixtureCombination()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-selectWeightedMixtureCombination"></a>}}
	
	\subsection{Method \code{selectWeightedMixtureCombination()}}{
		The mixture combination strategy for selecting a posterior \(R_t\) involves finding all the estimates that have a window that span the estimate date (this may involve longer windowed estimate from
dates in the future) and estimating a gamma distribution with the same mean and SD as the mixture of all the estimates. This combined estimate involves estimates of all different windows.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$selectWeightedMixtureCombination(weights)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{weights - the weighting of the different windows according to their length as a vector which much be the same length as the longest window, maxTau.}{ - (java expects a RNumericVector)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withDefaultPrior"></a>}}
	
	\subsection{Method \code{withDefaultPrior()}}{
		This sets the selection strategy for prior estimate of the reproduction number to be the fixed R0 value provided to the CoriEstimator::new constructor. All estimates use the same prior
and this is equivalent to the strategy implemented in EpiEstim (which has a default value of 5 for the mean and 5 for the SD of the R0 prior). This is an uninformed prior strategy.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$withDefaultPrior()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withFixedPrior"></a>}}
	
	\subsection{Method \code{withFixedPrior()}}{
		This sets the selection strategy for prior estimate of the reproduction number to be fixed value. All estimates use the same value as an uninformed prior, but enables us to compare the
impact of different priors.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$withFixedPrior(mean, sd)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{mean - the mean of the fixed prior \(R_t\)}{ - (java expects a double)}
				\item{sd - the sd  of the fixed prior \(R_t\)}{ - (java expects a double)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withAdaptivePrior"></a>}}
	
	\subsection{Method \code{withAdaptivePrior()}}{
		This sets the prior to be the posterior for the previous timestep, but with the SD of the previous posterior widened to allow the equivalent of a random walk. The widening factor defines how much 
change we expect to see in each time step
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$withAdaptivePrior(factor)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{factor - must be above 1 (or else the estimate will become locked on a specific value)}{ - (java expects a double)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-collectResampledQuantiles"></a>}}
	
	\subsection{Method \code{collectResampledQuantiles()}}{
		This configures how the estimator combines multiple estimates of \(R_t\) using different infectivity profiles. In this case the estimator will 
randomly sample from the selected posterior gamma distribution produced for each infectivity profile. These samples are combined and a mean, sd and empirical 
quantiles calculated.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$collectResampledQuantiles(sampleSize)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{sampleSize - the number of samples to draw for each posterior estimate of \(R_t\). This number will be multiplied by the number of infectivity profile estimates and the number time points, so should not be set too high or it will lead to memory consumption issues. This is the equivalent of the $n2$ parameter in EpiEstim}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-collectMixtureQuantiles"></a>}}
	
	\subsection{Method \code{collectMixtureQuantiles()}}{
		This configures how the estimator combines multiple estimates of \(R_t\) using different infectivity profiles. In this case it calculates a mixture distribution 
of all the posterior estimates, and combined values for mean, and standard deviation, of the mixture plus the quantiles using a brent solver.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$collectMixtureQuantiles()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-collectMixtureApproximation"></a>}}
	
	\subsection{Method \code{collectMixtureApproximation()}}{
		This configures how the estimator combines multiple estimates of \(R_t\) using different infectivity profiles. In this case it calculates a mixture distribution 
of all the posterior estimates, and then estimates a single gamma distribution with the same mean and SD as the mixture. This will only be a valid approximation
if all the posterior estimates are close to each other, which will be the case if the infectivitiy profiles are fairly similar. It shoudl tend to produce estimates
with less confidence. It is however the quickest method.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$collectMixtureApproximation()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-detailedOutput"></a>}}
	
	\subsection{Method \code{detailedOutput()}}{
		Configure the estimator to output full details for all infectivity profiles.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$detailedOutput()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withInfectivityProfile"></a>}}
	
	\subsection{Method \code{withInfectivityProfile()}}{
		Configure the estimator with an infectivity profile. At least one profile must be added
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$withInfectivityProfile(infectivityProfile, replace)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{infectivityProfile - A numeric vector of the daily discrete probability of a secondary infection event occurring given an infected individual. It is assumed that this vector starts 
at time zero with a probability of zero.}{ - (java expects a RNumericVector)}
				\item{replace -  replace the existing infectivity profile with this new one? if false then the profile is added to a list.}{ - (java expects a boolean)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withInfectivityProfileMatrix"></a>}}
	
	\subsection{Method \code{withInfectivityProfileMatrix()}}{
		Configure the estimator with an infectivity profile by specifying it as a matrix. This will replace all previously set profiles.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$withInfectivityProfileMatrix(infectivityProfiles)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{infectivityProfiles - A numeric array of the daily discrete probability of a secondary infection event occurring given an infected individual. It is assumed that this vector starts 
at time zero with a probability of zero.}{ - (java expects a RNumericArray)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-atStartOfTimeseries"></a>}}
	
	\subsection{Method \code{atStartOfTimeseries()}}{
		If the timeseries is at the start then we can safely impute zero values prior to the start of the timeseries. If the timeseries starts in the middle of the epidemic we must infer the 
infections prior to the beginning of the time series if we are to prevent sudden jumps in the estimates over the initial part of the timeseries.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$atStartOfTimeseries()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-inMiddleOfTimeseries"></a>}}
	
	\subsection{Method \code{inMiddleOfTimeseries()}}{
		If the timeseries starts in the middle of the epidemic we must infer the 
infections prior to the beginning of the time series if we are to prevent sudden jumps in the 
estimates over the initial part of the timeseries. This is done using the supplied value of R0 to 
infer a growth rate.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$inMiddleOfTimeseries()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-estimateRt"></a>}}
	
	\subsection{Method \code{estimateRt()}}{
		Estimates \(R_t\) for a one or more timeseries of infections in an epidemic, where the supplied dataframe is grouped
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$estimateRt(incidence, dateColName, incidenceColName)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{incidence - a dataframe containing at least 2 columns (one containing a date, the other a numeric count of cases). The dataframe may be grouped, in which case grouping is preserved, and each group is treated as a seperate time seriesincidenceColName - the name of the incidence column}{ - (java expects a RDataframe)}
				\item{dateColName - the name of the date column}{ - (java expects a String)}
				\item{incidenceColName - the name of the incidence column}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RDataframe: a dataframe, containing either a summary of the estimates of \(R_t\) with mean, sd, and quantiles for each day, or a full breakdown of the component \(R_t\) estimates.
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-estimateRtSingle"></a>}}
	
	\subsection{Method \code{estimateRtSingle()}}{
		Estimates \(R_t\) for a single timeseries of infections in an epidemic.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$estimateRtSingle(incidence, dateColName, incidenceColName)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{incidence - a dataframe containing at least 2 columns (one containing a date, the other a numeric count of cases). The dataframe may be grouped, in which case grouping is preservedincidenceColName - the name of the incidence column}{ - (java expects a RDataframe)}
				\item{dateColName - the name of the date column}{ - (java expects a String)}
				\item{incidenceColName - the name of the incidence column}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RDataframe: a dataframe, containing either a summary of the estimates of \(R_t\) with mean, sd, and quantiles for each day, or a full breakdown of the component \(R_t\) estimates.
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-estimateRtFromRates"></a>}}
	
	\subsection{Method \code{estimateRtFromRates()}}{
		Estimates \(R_t\) for a one or more timeseries of infections in an epidemic, where the supplied dataframe is grouped
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$estimateRtFromRates(rates, dateColName, rateColName, samplesPerProfile)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{rates - a dataframe containing at least 2 columns (one containing a date, the other a numeric poission rate of cases). The dataframe may be grouped, in which case grouping is preserved, and each group is treated as a seperate time series}{ - (java expects a RDataframe)}
				\item{dateColName - the name of the date column}{ - (java expects a String)}
				\item{rateColName - the name of the rate column}{ - (java expects a String)}
				\item{samplesPerProfile - the number of times to bootstrap per infectivity profile}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RDataframe: a dataframe, containing either a summary of the estimates of \(R_t\) with mean, sd, and quantiles for each day, or a full breakdown of the component \(R_t\) estimates.
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-estimateRtFromSingleRate"></a>}}
	
	\subsection{Method \code{estimateRtFromSingleRate()}}{
		Estimates \(R_t\) for a single timeseries of infections in an epidemic.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$CoriEstimator$new(...);
instance$estimateRtFromSingleRate(rate, dateColName, rateColName, samplesPerProfile)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{rate - a dataframe containing at least 2 columns (one containing a date, the other a numeric estimate of possion rate of cases). The dataframe may be grouped, in which case grouping is preservedrateColName - the name of the estimated poisson rate}{ - (java expects a RDataframe)}
				\item{dateColName - the name of the date column}{ - (java expects a String)}
				\item{rateColName - the name of the estimated poisson rate}{ - (java expects a String)}
				\item{samplesPerProfile - the number of times to bootstrap per infectivity profile}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RDataframe: a dataframe, containing either a summary of the estimates of \(R_t\) with mean, sd, and quantiles for each day, or a full breakdown of the component \(R_t\) estimates.
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
}