% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{DataframeSerialiser}

\alias{DataframeSerialiser}

\title{DataframeSerialiser}

\usage{\preformatted{
library(testRapi)
J = JavaApi$new();
instance <- J$DataframeSerialiser$new(...);
}}

\description{
	Actually useful class which takes an R dataframe and a filename
}

\details{
	
string and serialises the dataframe so it can be used natively in java
for testing purposes.
}

\examples{
## -----------------------------------
## Method `DataframeSerialiser$serialise`
## -----------------------------------

}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$DataframeSerialiser$new()}}
		}
	}
	\subsection{Static methods}{
		\itemize{
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-serialise}{\code{instance$serialise(dataframe, filename)}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$DataframeSerialiser$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 DataframeSerialiser object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-serialise"></a>}}
	
	\subsection{Method \code{serialise()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$DataframeSerialiser$new(...);
  instance$serialise(dataframe, filename)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{dataframe}{ - (java expects a uk.co.terminological.jsr233plugin.RType@d978ab9)}
				\item{filename}{ - (java expects a uk.co.terminological.jsr233plugin.RType@5a78b52b)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@7672960e: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}