% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-datasets.R
\name{discreteGammaInfectivityProfile}
\alias{discreteGammaInfectivityProfile}
\title{Generates one or more discrete infectivity profiles based on discretised gamma distributions.}
\usage{
discreteGammaInfectivityProfile(mean, sd, names = NULL, range = 20)
}
\arguments{
\item{mean}{- the mean(s) of the gamma distributions}

\item{sd}{- the sd(s) of the gamma distribution}

\item{names}{- the names of the infectivity profiles, if they represent a specific variant for example, if omitted they will be numbered}

\item{range}{- the number of days to generate - the first time point is always zero, so the resulting length will be one greater than this value.}
}
\value{
a list of class infectivity_profile containing aVector, the end points of the period, and yMatrix the discrete probabilities in the period up to the relevant aVector entry. Each column in yMatrix is a infectivity profile
}
\description{
Generates one or more discrete infectivity profiles based on discretised gamma distributions.
}
\examples{
discreteGammaInfectivityProfile(c(4,3),c(5,4),names=c("delta","omicron"),range=20)
}
