% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{doublingTimeFromGrowthRate}
\alias{doublingTimeFromGrowthRate}
\title{Calucualte a doubling time with quantiles for any timeseries with Growth rate estimates}
\usage{
doublingTimeFromGrowthRate(simpleTimeseries)
}
\arguments{
\item{simpleTimeseries}{}
}
\value{
a timeseries with doubling time estimates (columns starting with "doublingTime")
}
\description{
Calucualte a doubling time with quantiles for any timeseries with Growth rate estimates
}
