% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{epiestimRtEstimate}
\alias{epiestimRtEstimate}
\title{Minimal epiestim wrapper to execute a time series R_t using a discrete infectivity profile matrix, and format the result to be consistent with the rest of this..}
\usage{
epiestimRtEstimate(
  simpleTimeseries,
  yMatrix,
  bootstraps = 10 * dim(yMatrix)[2],
  window = 14
)
}
\description{
Minimal epiestim wrapper to execute a time series R_t using a discrete infectivity profile matrix, and format the result to be consistent with the rest of this..
}
