% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{FactoryTest}

\alias{FactoryTest}

\title{FactoryTest}

\usage{\preformatted{
library(testRapi)
J = JavaApi$new();
instance <- J$FactoryTest$new(...);
}}

\description{
	
}

\details{
	
}

\examples{
## -----------------------------------
## Method `FactoryTest$generateCharacter`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateNumeric`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateInteger`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateFactor`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateLogical`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateCharacterVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateNumericVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateIntegerVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateFactorVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateLogicalVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateDataframe`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateStreamDataFrame`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateList`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateNamedList`
## -----------------------------------

}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$FactoryTest$new()}}
		}
	}
	\subsection{Static methods}{
		\itemize{
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-generateCharacter}{\code{instance$generateCharacter()}}
			\item \href{#method-generateNumeric}{\code{instance$generateNumeric()}}
			\item \href{#method-generateInteger}{\code{instance$generateInteger()}}
			\item \href{#method-generateFactor}{\code{instance$generateFactor()}}
			\item \href{#method-generateLogical}{\code{instance$generateLogical()}}
			\item \href{#method-generateCharacterVec}{\code{instance$generateCharacterVec()}}
			\item \href{#method-generateNumericVec}{\code{instance$generateNumericVec()}}
			\item \href{#method-generateIntegerVec}{\code{instance$generateIntegerVec()}}
			\item \href{#method-generateFactorVec}{\code{instance$generateFactorVec()}}
			\item \href{#method-generateLogicalVec}{\code{instance$generateLogicalVec()}}
			\item \href{#method-generateDataframe}{\code{instance$generateDataframe()}}
			\item \href{#method-generateStreamDataFrame}{\code{instance$generateStreamDataFrame()}}
			\item \href{#method-generateList}{\code{instance$generateList()}}
			\item \href{#method-generateNamedList}{\code{instance$generateNamedList()}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$FactoryTest$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 FactoryTest object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateCharacter"></a>}}
	
	\subsection{Method \code{generateCharacter()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateCharacter()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@40013051: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateNumeric"></a>}}
	
	\subsection{Method \code{generateNumeric()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateNumeric()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@d180961: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateInteger"></a>}}
	
	\subsection{Method \code{generateInteger()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateInteger()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@5b59c3d: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateFactor"></a>}}
	
	\subsection{Method \code{generateFactor()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateFactor()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@6f65a203: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateLogical"></a>}}
	
	\subsection{Method \code{generateLogical()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateLogical()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@4de158ca: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateCharacterVec"></a>}}
	
	\subsection{Method \code{generateCharacterVec()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateCharacterVec()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@5f5a33ed: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateNumericVec"></a>}}
	
	\subsection{Method \code{generateNumericVec()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateNumericVec()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@4af7ac25: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateIntegerVec"></a>}}
	
	\subsection{Method \code{generateIntegerVec()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateIntegerVec()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@69e36fe3: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateFactorVec"></a>}}
	
	\subsection{Method \code{generateFactorVec()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateFactorVec()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@1d0b447b: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateLogicalVec"></a>}}
	
	\subsection{Method \code{generateLogicalVec()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateLogicalVec()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@4f2fdf3d: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateDataframe"></a>}}
	
	\subsection{Method \code{generateDataframe()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateDataframe()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@d978ab9: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateStreamDataFrame"></a>}}
	
	\subsection{Method \code{generateStreamDataFrame()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateStreamDataFrame()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@d978ab9: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateList"></a>}}
	
	\subsection{Method \code{generateList()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateList()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@3cf06a01: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateNamedList"></a>}}
	
	\subsection{Method \code{generateNamedList()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$FactoryTest$new(...);
  instance$generateNamedList()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@71ee2fff: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}