% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-datasets.R
\name{generateGrowthRate}
\alias{generateGrowthRate}
\title{Generate a growth timeseries simulation based on a small set of fixed time points.}
\usage{
generateGrowthRate(
  name,
  dateAtTime0 = "2020-01-01",
  length = 365,
  breaks = sort(c(0, sample(0:364, 4))),
  rates = alternatingRandom(),
  smooth = TRUE,
  sawtooth = FALSE,
  Gt.mean = 5,
  Gt.sd = 4,
  ...
)
}
\arguments{
\item{name}{- a meaningful name for this growth rate scenario}

\item{dateAtTime0}{- the date at time zero}

\item{length}{- how many days in the scenario}

\item{breaks}{- time points (between 0 and length) when the growth rate is at a set value.}

\item{rates}{- the values of the growth rate at the time points up to and including the break points, optionally an additional point for the period to the end of the time-series}

\item{smooth}{- should the rates be interpolated with a spline. If so it may be a good ideal to specify the growth rate at zero.}

\item{sawtooth}{- if the rates are not smooth should they be a step function or a saw-tooth linear function.}

\item{Gt.mean}{- mean of the generation time for conversion to R_t}

\item{Gt.sd}{- sd of the generation time for conversion to R_t}

\item{...}{- ignored}
}
\value{
- a list containing name - the scenario name,  ts - a tibble with the date, generated growth rate and R_t values, events - the time-points of the period, infectivityProfile - an infectivity_profile
}
\description{
Generate a growth timeseries simulation based on a small set of fixed time points.
}
