% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-datasets.R
\name{getGrowthRateBasedDataset}
\alias{getGrowthRateBasedDataset}
\title{Convenience method to generate a random growth rate simulation}
\usage{
getGrowthRateBasedDataset(
  growthSimulation = NULL,
  bootstraps = 100,
  seed = 100,
  weekendEffect = 0.1,
  Gt.mean = 5,
  Gt.sd = 4,
  periodic = FALSE,
  name = "synthetic",
  ...
)
}
\arguments{
\item{growthSimulation}{- optional growth rate timeseries, if missing a random one is generated}

\item{bootstraps}{- ho many replicates}

\item{seed}{- what initial case loads}

\item{weekendEffect}{- any weekend effect?}

\item{Gt.mean}{- generation time mean}

\item{Gt.sd}{- generation time sd}

\item{periodic}{- a random periodic or}

\item{name}{- a name for the simulation}

\item{...}{}
}
\value{
a single strain random simulation with sensible parameters and no observation delay
}
\description{
Convenience method to generate a random growth rate simulation
}
