% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-datasets.R
\name{getTwoAlternativesDataset}
\alias{getTwoAlternativesDataset}
\title{Convenience wrapper to generate a 2 strain type comparison. strain 1 will have suffix .s1 and strain 2 will have .s2}
\usage{
getTwoAlternativesDataset(
  scenario1,
  scenario2,
  delayMean1,
  delayMean2,
  delaySd1 = delayMean1/4,
  delaySd2 = delayMean2/4,
  timepoints = nrow(scenario1$ts),
  ...
)
}
\arguments{
\item{scenario1}{a growthSimulation scenario for the s1 strain}

\item{scenario2}{a growthSimulation scenario for the s2 strain}

\item{delayMean1}{a delay distribution parameter for the s1 strain}

\item{delayMean2}{a delay distribution parameter for the s2 strain}

\item{delaySd1}{a delay distribution parameter for the s1 strain}

\item{delaySd2}{a delay distribution parameter for the s2 strain}

\item{timepoints}{times at which to generate observed time series.}

\item{...}{}
}
\value{
a 2 strain dataset in wide format.
}
\description{
Convenience wrapper to generate a 2 strain type comparison. strain 1 will have suffix .s1 and strain 2 will have .s2
}
