% Generated by r6-generator-maven-plugin: do not edit by hand

\name{GrowthRateEstimator}

\alias{GrowthRateEstimator}

\title{GrowthRateEstimator}

\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{minWindow}{minWindow - the smallest amount of data considered in estimating the growth rate - 4 is a ssensible default unless you want bang up to date estimate.minWindow - the largest amount of data considered in estimating the growth rate - 14 is a sensible amount . - (java expects a int)}
	\item{maxWindow}{maxWindow - (java expects a int)}
	\if{html}{\out{</div>}}
}


\description{
An estimator of the reproduction number (\(R_t\)) using a renewal equation method. 

Version: 0.03

Generated: 2021-12-28T11:33:45.875915
}

\details{
	
The renewal equation method depends on a time series of infections, and on the infectivity profile - 
a measure of the probability that a secondary infection occurred on a specific day after the primary case, 
given a secondary infection occurred. A Bayesian framework is then used to update a prior probabilistic 
estimate of \(R_t\) on any given day with both information gained from the time series of infections in the 
epidemic to date and the infectivity profile to produce a posterior estimate of \(R_t\). This is based on the 
article:<br>

<br>A. GrowthRate, N. M. Ferguson, C. Fraser, and S. Cauchemez, ‘A New Framework and Software to Estimate Time-Varying Reproduction Numbers During Epidemics’, Am J Epidemiol, vol. 178, no. 9, pp. 1505–1512, Nov. 2013, doi: 10.1093/aje/kwt133.<br>

<br> and the implementation of it in the R Library EpiEstim. It assumes that infections (or other type of observation) are a 
Poisson distributed process and that the reproduction number is a gamma distributed quantity. The estimates of \(R_t\) may be 
aggregated over a specific "window" to provide a more stable estimate.
}

\examples{\dontrun{
J = jepidemic::JavaApi$get();
instance = J$GrowthRateEstimator$new(minWindow, maxWindow)

## -----------------------------------
## Method `GrowthRateEstimator$strictGrowthRateAndRtEstimator`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$defaultGrowthRateAndRtEstimator`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$defaultGrowthRateEstimator`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$basicGrowthRateEstimator`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$withOutputQuantiles`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$withSaneDefaults`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$withInitialIncidence`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$withGaussianEstimateWeighting`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$withEstimateWeighting`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$combineEstimatesWithWeightedMixture`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$priorIncidenceFromPreviousPosterior`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$priorIncidenceFromScaledPreviousPosterior`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$priorIncidenceFromPosteriorAndGrowthRate`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$priorIncidenceFromScaledPosteriorAndGrowthRate`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$useAllPosteriorEstimates`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$usePosteriorEstimatesFromOneWindow`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$usePosteriorEstimatesWithEnoughData`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$withInfectivityProfile`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$withInfectivityProfileMatrix`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$estimateGrowthRate`
## -----------------------------------

## -----------------------------------
## Method `GrowthRateEstimator$estimateGrowthRateSingle`
## -----------------------------------

}}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$GrowthRateEstimator$new(minWindow, maxWindow)}}
		}
	}
	\subsection{Static methods}{
		\itemize{
			\item \href{#method-strictGrowthRateAndRtEstimator}{\code{J$GrowthRateEstimator$strictGrowthRateAndRtEstimator(minWindow, maxWindow, infectivityProfiles)}}
			\item \href{#method-defaultGrowthRateAndRtEstimator}{\code{J$GrowthRateEstimator$defaultGrowthRateAndRtEstimator(minWindow, maxWindow, initialIncidence)}}
			\item \href{#method-defaultGrowthRateEstimator}{\code{J$GrowthRateEstimator$defaultGrowthRateEstimator(minWindow, maxWindow)}}
			\item \href{#method-basicGrowthRateEstimator}{\code{J$GrowthRateEstimator$basicGrowthRateEstimator()}}
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-withOutputQuantiles}{\code{instance$withOutputQuantiles(quantiles)}}
			\item \href{#method-withSaneDefaults}{\code{instance$withSaneDefaults()}}
			\item \href{#method-withInitialIncidence}{\code{instance$withInitialIncidence(incidence)}}
			\item \href{#method-withGaussianEstimateWeighting}{\code{instance$withGaussianEstimateWeighting(meanTau, sdTau)}}
			\item \href{#method-withEstimateWeighting}{\code{instance$withEstimateWeighting(weights)}}
			\item \href{#method-combineEstimatesWithWeightedMixture}{\code{instance$combineEstimatesWithWeightedMixture()}}
			\item \href{#method-priorIncidenceFromPreviousPosterior}{\code{instance$priorIncidenceFromPreviousPosterior(multiplyingSDby)}}
			\item \href{#method-priorIncidenceFromScaledPreviousPosterior}{\code{instance$priorIncidenceFromScaledPreviousPosterior()}}
			\item \href{#method-priorIncidenceFromPosteriorAndGrowthRate}{\code{instance$priorIncidenceFromPosteriorAndGrowthRate(multiplyingSDby)}}
			\item \href{#method-priorIncidenceFromScaledPosteriorAndGrowthRate}{\code{instance$priorIncidenceFromScaledPosteriorAndGrowthRate()}}
			\item \href{#method-useAllPosteriorEstimates}{\code{instance$useAllPosteriorEstimates()}}
			\item \href{#method-usePosteriorEstimatesFromOneWindow}{\code{instance$usePosteriorEstimatesFromOneWindow(tau)}}
			\item \href{#method-usePosteriorEstimatesWithEnoughData}{\code{instance$usePosteriorEstimatesWithEnoughData(sumIncidence)}}
			\item \href{#method-withInfectivityProfile}{\code{instance$withInfectivityProfile(infectivityProfile, replace)}}
			\item \href{#method-withInfectivityProfileMatrix}{\code{instance$withInfectivityProfileMatrix(infectivityProfiles)}}
			\item \href{#method-estimateGrowthRate}{\code{instance$estimateGrowthRate(incidence, dateColName, incidenceColName)}}
			\item \href{#method-estimateGrowthRateSingle}{\code{instance$estimateGrowthRateSingle(incidence, dateColName, incidenceColName)}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		Produces a GrowthRateEstimator that is configured to behave in a sensible fashion, with sane defaults. 
In that it estimates an initial poisson rate prior from the first entry and returns a fixed window size and aggregates uncertain SI 
distributions with a random resampling strategy<br><br>
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
J$GrowthRateEstimator$new(minWindow, maxWindow)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{minWindow - the smallest amount of data considered in estimating the growth rate - 4 is a ssensible default unless you want bang up to date estimate.minWindow - the largest amount of data considered in estimating the growth rate - 14 is a sensible amount .}{ - (java expects a int)}
				\item{maxWindow}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a default growth rate estimator
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-strictGrowthRateAndRtEstimator"></a>}}
	
	\subsection{Method \code{strictGrowthRateAndRtEstimator()}}{
		Produces a GrowthRateEstimator that will estimate growth rate and $R_t$, with sane defaults. The $R_t$ estimation is 
done using bootstrapping, and combining the results using a mixture distribution. This will be relatively slow.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
J$GrowthRateEstimator$strictGrowthRateAndRtEstimator(minWindow, maxWindow, infectivityProfiles)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{minWindow - the smallest amount of data considered in estimating the growth rate - 4 is a ssensible default unless you want bang up to date estimate.minWindow - the largest amount of data considered in estimating the growth rate - 14 is a sensible amount.}{ - (java expects a int)}
				\item{maxWindow}{ - (java expects a int)}
				\item{infectivityProfiles - a numeric matrix with each column containing a discrete probability distribution of the serial interval with \(P_{(t=0)} = 0\)}{ - (java expects a RNumericArray)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a strict growth rate and $R_t$ estimator that accounts for uncertainty in the infectivity profile.
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-defaultGrowthRateAndRtEstimator"></a>}}
	
	\subsection{Method \code{defaultGrowthRateAndRtEstimator()}}{
		Produces a GrowthRateEstimator that will estimate growth rate and $R_t$, with sane defaults. The $R_t$ estimation is 
done using a single infectivity profile and does not account for infectivity profile uncertainty. This is comparatively fast.<br><br>
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
J$GrowthRateEstimator$defaultGrowthRateAndRtEstimator(minWindow, maxWindow, initialIncidence)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{minWindow - the smallest amount of data considered in estimating the growth rate - 4 is a ssensible default unless you want bang up to date estimate.minWindow - the largest amount of data considered in estimating the growth rate - 14 is a sensible amount.}{ - (java expects a int)}
				\item{maxWindow}{ - (java expects a int)}
				\item{initialIncidence}{ - (java expects a double)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a growth rate and $R_t$ estimator that does not account for uncertainty in the infectivity profile.
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-defaultGrowthRateEstimator"></a>}}
	
	\subsection{Method \code{defaultGrowthRateEstimator()}}{
		Produces a GrowthRateEstimator that will estimate growth rate only, with sane defaults, unless an infectivity profile is lated added<br><br>
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
J$GrowthRateEstimator$defaultGrowthRateEstimator(minWindow, maxWindow)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{minWindow - the smallest amount of data considered in estimating the growth rate - 4 is a ssensible default unless you want bang up to date estimate.minWindow - the largest amount of data considered in estimating the growth rate - 14 is a sensible amount.}{ - (java expects a int)}
				\item{maxWindow}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a growth rate and $R_t$ estimator that does not account for uncertainty in the infectivity profile.
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-basicGrowthRateEstimator"></a>}}
	
	\subsection{Method \code{basicGrowthRateEstimator()}}{
		Produces a GrowthRateEstimator that will estimate growth rate only, with sane defaults, and estimating using windows between 4 and 14 days<br><br>
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
J$GrowthRateEstimator$basicGrowthRateEstimator()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a growth rate and $R_t$ estimator that does not account for uncertainty in the infectivity profile.
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withOutputQuantiles"></a>}}
	
	\subsection{Method \code{withOutputQuantiles()}}{
		Allows you to change the quantiles returned by the estimator. The defaults are c(0.025,0.05,0.25,0.5,0.75,0.95,0.975)
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$withOutputQuantiles(quantiles)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{quantiles the desired quantiles of the output}{ - (java expects a RNumericVector)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withSaneDefaults"></a>}}
	
	\subsection{Method \code{withSaneDefaults()}}{
		Provides a sensible baseline configuration for the estimator that involves estimating the poisson rate using the 
previous timestep's posterior with standard deviation scaled by a factor determined by the t-2 value. Multiple posterior
estimates are combined using a weighted average based on a discretised normal distributed weight with mean of 7 and SD of 4.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$withSaneDefaults()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withInitialIncidence"></a>}}
	
	\subsection{Method \code{withInitialIncidence()}}{
		Allows configuration of the initial incidence. If this is omitted the initial incidence is guessed from the first entry in the timeseries
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$withInitialIncidence(incidence)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{incidence}{ - (java expects a double)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withGaussianEstimateWeighting"></a>}}
	
	\subsection{Method \code{withGaussianEstimateWeighting()}}{
		Allows configuration of the weighting of estimate combination for the weighted mixture combination. 
This uses a discretised normal distribution to generate values for the weighting. The weighting is 
applied to each supported estimate determined by the data window of that estimate.  The default value
is a mean of 7 and SD of 4
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$withGaussianEstimateWeighting(meanTau, sdTau)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{meanTau the mean of the normal distribution of weights}{ - (java expects a double)}
				\item{sdTau the sd of the normal distribution of weights}{ - (java expects a double)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withEstimateWeighting"></a>}}
	
	\subsection{Method \code{withEstimateWeighting()}}{
		Allows configuration of the weighting of estimate combination for the weighted mixture combination. 
This assumes the weights are given as a vector where the first entry is the weight for an estimate using a 
data window of 1. The vector should be the same length as the maxWindow parameter used in configuring the 
GrowthRateEstimator or 14 if the defaults were used.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$withEstimateWeighting(weights)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{weights - a vector of weights for each data window}{ - (java expects a RNumericVector)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-combineEstimatesWithWeightedMixture"></a>}}
	
	\subsection{Method \code{combineEstimatesWithWeightedMixture()}}{
		Sets the estimator to use a weighted mixture to combine estimates based on the length of different data windows, using weights defined 
by withEstimateWeighting() or withGaussianEstimateWeighting(). If none is set the weighting defaults to a discretised normal
distribution with mean of 7 an SD of 4.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$combineEstimatesWithWeightedMixture()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-priorIncidenceFromPreviousPosterior"></a>}}
	
	\subsection{Method \code{priorIncidenceFromPreviousPosterior()}}{
		Select a prior estimate for the poisson rate by taking the previous posterior and multiplying 
the SD by a constant factor
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$priorIncidenceFromPreviousPosterior(multiplyingSDby)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{multiplyingSDby - factor to multiply SD by. Must be greater than or equal to 1.}{ - (java expects a double)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-priorIncidenceFromScaledPreviousPosterior"></a>}}
	
	\subsection{Method \code{priorIncidenceFromScaledPreviousPosterior()}}{
		Select a prior estimate for the poisson rate by taking the previous posterior and multiplying 
the SD by a factor determined by the difference between the t-1 and t-2 incidence. This allows for 
rapid rate of change to be detected and less weight given to the prior in this situation
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$priorIncidenceFromScaledPreviousPosterior()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-priorIncidenceFromPosteriorAndGrowthRate"></a>}}
	
	\subsection{Method \code{priorIncidenceFromPosteriorAndGrowthRate()}}{
		Select a prior estimate for the poisson rate by taking the previous posterior and increasing it by  
the exponential previous posterior growth rate and multiplying the resulting disctribution by a 
constant factor. This is a mechanistic prior.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$priorIncidenceFromPosteriorAndGrowthRate(multiplyingSDby)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{multiplyingSDby - factor to multiply SD by. Must be greater than or equal to 1.}{ - (java expects a double)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-priorIncidenceFromScaledPosteriorAndGrowthRate"></a>}}
	
	\subsection{Method \code{priorIncidenceFromScaledPosteriorAndGrowthRate()}}{
		Select a prior estimate for the poisson rate by taking the previous posterior and increasing it by  
the exponential previous posterior growth rate and multiplying the resulting disctribution by a 
factor determined by the change in t-1 and t-2 indidence. This is a mechanistic prior.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$priorIncidenceFromScaledPosteriorAndGrowthRate()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-useAllPosteriorEstimates"></a>}}
	
	\subsection{Method \code{useAllPosteriorEstimates()}}{
		The posterior estimates for each combination of data window and infectivity profile are all retained and combined
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$useAllPosteriorEstimates()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-usePosteriorEstimatesFromOneWindow"></a>}}
	
	\subsection{Method \code{usePosteriorEstimatesFromOneWindow()}}{
		The posterior estimates for only one data window, but all infectivity profiles, are retained and combined. N.b. the same effect
can be achieved by setting the minWindow and maxWindow in the constructor.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$usePosteriorEstimatesFromOneWindow(tau)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{tau the window width to retain.}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-usePosteriorEstimatesWithEnoughData"></a>}}
	
	\subsection{Method \code{usePosteriorEstimatesWithEnoughData()}}{
		The posterior estimates for data windows that include a reasonable amount of data are retained and combined. This 
suppresses noisy estimates made on a very few data points when incidence is low.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$usePosteriorEstimatesWithEnoughData(sumIncidence)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{sumIncidence}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withInfectivityProfile"></a>}}
	
	\subsection{Method \code{withInfectivityProfile()}}{
		Configure the estimator with an infectivity profile. At least one profile must be added for an estimate of R_t to be made
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$withInfectivityProfile(infectivityProfile, replace)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{infectivityProfile - A numeric vector of the daily discrete probability of a secondary infection event occurring given an infected individual. It is assumed that this vector starts 
at time zero with a probability of zero.}{ - (java expects a RNumericVector)}
				\item{replace -  replace the existing infectivity profile with this new one? if false then the profile is added to a list.}{ - (java expects a boolean)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withInfectivityProfileMatrix"></a>}}
	
	\subsection{Method \code{withInfectivityProfileMatrix()}}{
		Configure the estimator with an infectivity profile by specifying it as a matrix. This will replace all previously set profiles.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$withInfectivityProfileMatrix(infectivityProfiles)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{infectivityProfiles - A numeric array of the daily discrete probability of a secondary infection event occurring given an infected individual in each column. It is assumed that this vector starts 
at time zero with a probability of zero.}{ - (java expects a RNumericArray)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 GrowthRateEstimator object: The estimator itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-estimateGrowthRate"></a>}}
	
	\subsection{Method \code{estimateGrowthRate()}}{
		Estimates growth rate and, if configured with an infectivity profile, \(R_t\) for a one or more timeseries of infections in an epidemic. If the supplied dataframe is grouped it assumes ach grouping is a timeseries
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$estimateGrowthRate(incidence, dateColName, incidenceColName)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{incidence - a dataframe containing at least 2 columns (one containing a date, the other a numeric count of cases). The dataframe may be grouped, in which case grouping is preserved, and each group is treated as a seperate time seriesincidenceColName - the name of the incidence column}{ - (java expects a RDataframe)}
				\item{dateColName - the name of the date column}{ - (java expects a String)}
				\item{incidenceColName - the name of the incidence column}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RDataframe: a dataframe, containing a summary of the estimates of incidence, growth rate and \(R_t\) with mean, sd, and quantiles for each day.
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-estimateGrowthRateSingle"></a>}}
	
	\subsection{Method \code{estimateGrowthRateSingle()}}{
		Estimates \(R_t\) for a single timeseries of infections in an epidemic.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
instance = J$GrowthRateEstimator$new(...);
instance$estimateGrowthRateSingle(incidence, dateColName, incidenceColName)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{incidence - a dataframe containing at least 2 columns (one containing a date, the other a numeric count of cases).incidenceColName - the name of the incidence column}{ - (java expects a RDataframe)}
				\item{dateColName - the name of the date column}{ - (java expects a String)}
				\item{incidenceColName - the name of the incidence column}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RDataframe: a dataframe, containing either a summary of the estimates of \(R_t\) with mean, sd, and quantiles for each day, or a full breakdown of the component \(R_t\) estimates.
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
}