% Generated by r6-generator-maven-plugin: do not edit by hand
\docType{package}
\name{html2pdfr-package}
\alias{html2pdfr}
\alias{html2pdfr-package}
\title{html2pdfr: R wrapper for OpenHTMLtoPDF java library}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}
This library is really designed to help get PDF or PNG formats of HTML tables, with layout determined by page size and control of styling determined by CSS. This allows use of HTML table layout within latex documents. Although tabular content should be allowed to grow up to the page dimensions, we don't want to force table layout to be a set width, but preventing it from overflowing bounding boxes is also essential. Although the focus is on tables, any basic HTML content can be rendered including SVG and MathML, up to the support of the underlying HTML rendering engine (https://github.com/danfickle/openhtmltopdf). This does not for example include javascript which would be required for D3 content.

Version: 0.01

Classes:
\itemize{ 
	\item \code{JavaApi}
	\item \code{HtmlConverter}
}
}

\author{
\itemize{ 
   \item Rob Challen rob.challen@bristol.ac.uk 0000-0002-5504-7768
}
}

\keyword{java api}