% Generated by r6-generator-maven-plugin: do not edit by hand

\name{InfectivityProfile}

\alias{InfectivityProfile}

\title{InfectivityProfile}

\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{discretePdf}{discretePdf - A discrete pdf of probabilities over time of secondary infection given priamry infection where probability at time 0 is 0 - (java expects a RNumericVector)}
	\item{id}{id - a numeric index or id for this profile - (java expects a int)}
	\if{html}{\out{</div>}}
}


\description{
A holder of a discrete probability distribution of the infectivity profile

Version: 0.03

Generated: 2022-05-24T22:12:30.975050
}

\details{
	A holder of a discrete probability distribution of the infectivity profile
}

\examples{\dontrun{
J = jepidemic::JavaApi$get();
instance = J$InfectivityProfile$new(discretePdf, id)

}}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$InfectivityProfile$new(discretePdf, id)}}
		}
	}
	\subsection{Static methods}{
		\itemize{
			\item{none}
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item{none}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
J$InfectivityProfile$new(discretePdf, id)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{discretePdf - A discrete pdf of probabilities over time of secondary infection given priamry infection where probability at time 0 is 0}{ - (java expects a RNumericVector)}
				\item{id - a numeric index or id for this profile}{ - (java expects a int)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 InfectivityProfile object: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
}