% Generated by r6-generator-maven-plugin: do not edit by hand

\name{JavaApi}

\alias{JavaApi}

\title{Java Epidemic}

\section{Usage}{
	\if{html}{\out{<div class="r">}}
	\preformatted{
 J = jepidemic::JavaApi$get(logLevel)
 	}
  \if{html}{\out{</div>}}
}

\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{logLevel}{optional - the slf4j log level as a string - one of OFF (most specific, no logging), 
	FATAL (most specific, little data), ERROR, WARN, INFO, DEBUG, 
	TRACE (least specific, a lot of data), ALL (least specific, all data)}
	\if{html}{\out{</div>}}
}

\description{
Some work in progress R functions related to the covid 19 epidemic

Version: 0.03

Generated: 2021-11-06T14:23:49.321748
}

\author{\email{rc538@exeter.ac.uk}}

\keyword{java api}

\section{Static methods and constructors}{
	\itemize{
		\item \code{JavaApi$get()}
		\item \href{#method-InfectivityProfile-new}{\code{J$InfectivityProfile$new(discretePdf, id)}}
		\item \href{#method-CoriEstimator-new}{\code{J$CoriEstimator$new(r0Mean, r0SD, maxWindow)}}
		\item \href{#method-CoriEstimator-defaultEpiEstim}{\code{J$CoriEstimator$defaultEpiEstim(infectivityProfile, meanR0, sdR0, window)}}
		\item \href{#method-CoriEstimator-defaultUncertainEpiEstim}{\code{J$CoriEstimator$defaultUncertainEpiEstim(infectivityProfiles, meanR0, sdR0, window, n2)}}
	}
	

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-InfectivityProfile-new"></a>}}
	
	\subsection{Method \code{InfectivityProfile$new()}}{
		no description
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
J$InfectivityProfile$new(discretePdf, id)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{discretePdf - A discrete pdf of probabilities over time of secondary infection given priamry infection where probability at time 0 is 0}{ - (java expects a RNumericVector)}
					\item{id - a numeric index or id for this profile}{ - (java expects a int)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 InfectivityProfile object: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-CoriEstimator-new"></a>}}
	
	\subsection{Method \code{CoriEstimator$new()}}{
		The Estimator is configured with the baseline \(R_0\) assumptions which are used as prior values for \(R_t\) when no other 
values are available, or the {@link CoriEstimator#withDefaultPrior()} option is used.
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
J$CoriEstimator$new(r0Mean, r0SD, maxWindow)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{r0Mean - The mean of the gamma distribution of the baseline prior assumption about R0 - the theorerical reproduction number in an infinite pool of evenly mixing susceptible individuals}{ - (java expects a double)}
					\item{r0SD - The SD of the prior gamma distribution}{ - (java expects a double)}
					\item{maxWindow - The maximum sliding window over which to estimate \(R_t\).}{ - (java expects a int)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-CoriEstimator-defaultEpiEstim"></a>}}
	
	\subsection{Method \code{CoriEstimator$defaultEpiEstim()}}{
		Produces a CoriEstimator that is configured to behave like the default settings of EpiEstim. 
In that it uses a fixed prior for Rt and returns a fixed window size<br><br>
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
J$CoriEstimator$defaultEpiEstim(infectivityProfile, meanR0, sdR0, window)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{infectivityProfile - a vector with the discrete probability distribution of the serial interval with \(P_{(t=0)} = 0\)}{ - (java expects a RNumericVector)}
					\item{meanR0 - the mean prior of R0}{ - (java expects a double)}
					\item{sdR0 - the sd prior of R0}{ - (java expects a double)}
					\item{window - the window size}{ - (java expects a int)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: a default cori estimator
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-CoriEstimator-defaultUncertainEpiEstim"></a>}}
	
	\subsection{Method \code{CoriEstimator$defaultUncertainEpiEstim()}}{
		Produces a CoriEstimator that is configured to behave like the default settings of EpiEstim. 
In that it uses a fixed prior for Rt and returns a fixed window size and aggregates uncertain SI 
distributions with a random resampling strategy<br><br>
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = jepidemic::JavaApi$get()
J$CoriEstimator$defaultUncertainEpiEstim(infectivityProfiles, meanR0, sdR0, window, n2)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{infectivityProfiles - a numeric matrix with the discrete probability distribution of the serial interval with \(P_{(t=0)} = 0\)}{ - (java expects a RNumericArray)}
					\item{meanR0 - the mean prior of R0}{ - (java expects a double)}
					\item{sdR0 - the sd prior of R0}{ - (java expects a double)}
					\item{window - the window size}{ - (java expects a int)}
					\item{n2 - the number of samples to use for estimating confidence limits on bootstrapped infectivity profiles}{ - (java expects a int)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 CoriEstimator object: a default cori estimator
		}
	}

}