% Generated by r6-generator-maven-plugin: do not edit by hand
\docType{package}
\name{jepidemic-package}
\alias{jepidemic}
\alias{jepidemic-package}
\title{jepidemic: Java Epidemic}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}
Some work in progress R functions related to the covid 19 epidemic

Version: 0.03

Classes:
\itemize{ 
	\item \code{JavaApi}
	\item \code{GrowthRateEstimator}
	\item \code{Serialiser}
	\item \code{InfectivityProfile}
	\item \code{CoriEstimator}
}
}

\author{
\itemize{ 
   \item Rob Challen rc538@exeter.ac.uk 0000-0002-5504-7768
   \item terminological   
}
}

\keyword{java api}