% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{locfitPoissonRateEstimate}
\alias{locfitPoissonRateEstimate}
\title{Generate a smoothed estimate of the absolute growth rate of cases using a poisson model.}
\usage{
locfitPoissonRateEstimate(
  simpleTimeseries,
  degree = 2,
  window = 14,
  weightByWeekday = FALSE,
  ...
)
}
\arguments{
\item{simpleTimeseries}{- a minimal time-series including date, value, and if available total. If total is present the proportion is value/total. otherwise it is value.}

\item{degree}{the polynomial degree}

\item{window}{the data window in days}

\item{...}{may include "nearestNeigbour=FALSE" to disable the tail behaviour of locfit}
}
\value{
a timeseries with poisson rate estimates (columns starting with "Est")
}
\description{
Generate a smoothed estimate of the absolute growth rate of cases using a poisson model.
}
