% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{locfitProportionalGrowthEstimate}
\alias{locfitProportionalGrowthEstimate}
\title{Generate a smoothed estimate of the relative growth rate of cases compared to some baseline using proportions.}
\usage{
locfitProportionalGrowthEstimate(
  simpleTimeseries,
  degree = 2,
  window = 14,
  ...
)
}
\arguments{
\item{simpleTimeseries}{- a minimal time-series including date, value, and if available total. If total is present the proportion is value/total. otherwise it is value, and total is assumed to be 1.}

\item{degree}{the polynomial degree}

\item{window}{the data window in days}

\item{...}{may include "nearestNeigbour=FALSE" to disable the tail behaviour of locfit}
}
\value{
a timeseries with growth rate estimates (columns starting with "Growth")
}
\description{
Generate a smoothed estimate of the relative growth rate of cases compared to some baseline using proportions.
}
