% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{locfitProportionEstimate}
\alias{locfitProportionEstimate}
\title{Generate a smoothed estimate of the proportion of cases compared to some total.}
\usage{
locfitProportionEstimate(
  simpleTimeseries,
  degree = 2,
  window = 14,
  estimateMean = FALSE,
  ...
)
}
\arguments{
\item{simpleTimeseries}{- a minimal time-series including date, value, and if available total. If total is present the proportion is value/total. otherwise it is value.}

\item{degree}{the polynomial degree}

\item{window}{the data window in days}

\item{estimateMean}{there is no closed form estimate of the mean of a logit transformed normal. it can be calculated by integration by this is relatively expensive and not done unless explicitly needed,}

\item{...}{may include "nearestNeigbour=FALSE" to disable the tail behaviour of locfit}
}
\value{
a timeseries with binomial proportion estimates (columns starting with "Proportion")
}
\description{
Generate a smoothed estimate of the proportion of cases compared to some total.
}
