% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{MinimalExample}

\alias{MinimalExample}

\title{MinimalExample}

\usage{\preformatted{
library(testRapi)
J = JavaApi$new();
instance <- J$MinimalExample$new(...);
}}

\description{
	This class is a very basic example of the features of the rJava maven plugin. <br/>
}

\details{
	

The class is annotated with an @RClass to identify it as part of the R API. <br/>

R code and documentation will be generated to allow the use of these functions in R through 
R6 Classes.
}

\examples{
## -----------------------------------
## Method `MinimalExample$demo`
## -----------------------------------
  J = testRapi::JavaApi$new()
  minExample = J$MinimalExample$new()
  dfTest = tibble::tibble(input=c(1,2,3))
  minExample$demo(dataframe=dfTest,message='Hello world')

## -----------------------------------
## Method `MinimalExample$demoStatic`
## -----------------------------------
  J = testRapi::JavaApi$new()
  J$MinimalExample$demoStatic('Ola, el mundo')

}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$MinimalExample$new(logMessage)}}
		}
	}
	\subsection{Static methods}{
		\itemize{
			\item \href{#method-demoStatic}{\code{J$MinimalExample$demoStatic(message)}}
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-demo}{\code{instance$demo(dataframe, message)}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		A maximum of one constructor of any signature can be used. </br>

If different constructors are needed then they may be used but not 
included in the R Api (i.e. not annotated with @RMethod. </br>

Static factory methods can be used instead.
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$MinimalExample$new(logMessage)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{logMessage - a message which will be logged}{ - (java expects a uk.co.terminological.jsr233plugin.RType@5a78b52b)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 MinimalExample object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
    
		J = testRapi::JavaApi$new()
  minExample = J$MinimalExample$new('initialising....')
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-demo"></a>}}
	
	\subsection{Method \code{demo()}}{
		The majority of the documentation of the method can be done in JavaDoc
and these will be reflected in the R documentation
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$MinimalExample$new(...);
  instance$demo(dataframe, message)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{dataframe - a dataframe with an arbitrary number of columns}{ - (java expects a uk.co.terminological.jsr233plugin.RType@d978ab9)}
				\item{message - a message}{ - (java expects a uk.co.terminological.jsr233plugin.RType@5a78b52b)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@d978ab9: the dataframe unchanged
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
  J = testRapi::JavaApi$new()
  minExample = J$MinimalExample$new()
  dfTest = tibble::tibble(input=c(1,2,3))
  minExample$demo(dataframe=dfTest,message='Hello world')
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-demoStatic"></a>}}
	
	\subsection{Method \code{demoStatic()}}{
		Static methods are also supported. These are accessed through the
root of the R api.
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$MinimalExample$demoStatic(message)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message}{ - (java expects a uk.co.terminological.jsr233plugin.RType@5a78b52b)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@7672960e: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
  J = testRapi::JavaApi$new()
  J$MinimalExample$demoStatic('Ola, el mundo')
		}
		\if{html}{\out{</div>}}
	}
}