% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-datasets.R
\name{periodicGrowthRate}
\alias{periodicGrowthRate}
\title{Generate a periodic growth timeseries simulation}
\usage{
periodicGrowthRate(
  name,
  dateAtTime0 = "2020-01-01",
  length = 365,
  period = 91,
  baseline = 0,
  smooth = TRUE,
  amplitude = 0.1,
  Gt.mean = 5,
  Gt.sd = 4
)
}
\arguments{
\item{name}{- a meaningful name for this growth rate scenario}

\item{dateAtTime0}{- the date at time zero}

\item{length}{- how many days in the scenario}

\item{period}{- the period of the periodic function}

\item{baseline}{- a baseline growth rate around which to oscillate usually 0.}

\item{smooth}{- smooth (sinusoidal) or triangular wave.}

\item{amplitude}{- growth rate (or R_t) amplitude}

\item{Gt.mean}{- mean of the generation time for conversion to R_t}

\item{Gt.sd}{- sd of the generation time for conversion to R_t}
}
\value{
- a list containing name - the scenario name,  ts - a tibble with the date, generated growth rate and R_t values, events - the time-points of the period, infectivityProfile - an infectivity_profile
}
\description{
Generate a periodic growth timeseries simulation
}
