% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{pointPoissonEstimate}
\alias{pointPoissonEstimate}
\title{Calculate a slightly more robust estimate of growth rate and  poisson rate based on a single model}
\usage{
pointPoissonEstimate(
  simpleTimeseries,
  dates,
  window,
  weekly = "weekday",
  includeModel = TRUE,
  ...
)
}
\arguments{
\item{simpleTimeseries}{- the timeseries containing date and value}

\item{dates}{- dates at which to evaluate the model}

\item{window}{- the window of data}

\item{weekly}{- either "weekday","weekend" or "none" to define whether to fit a fixed effect model to the weekday, or the is.weekend}

\item{includeModel}{- keep the fitted model as a list column for fit analysis}

\item{...}{}
}
\value{
a dataframe of evaluation dates, growth rates, poisson rates and model fit
}
\description{
Calculate a slightly more robust estimate of growth rate and  poisson rate based on a single model
}
