% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{pointProportionEstimate}
\alias{pointProportionEstimate}
\title{Calculate a slightly more robust estimate of growth rate and proportion based on a single model at a range of}
\usage{
pointProportionEstimate(
  simpleTimeseries,
  dates = max(simpleTimeseries$date) - 3,
  window = 14,
  weekly = "weekday",
  includeModel = TRUE,
  ...
)
}
\arguments{
\item{simpleTimeseries}{- the timeseries containing date, value and total}

\item{dates}{- dates at which to evaluate the model}

\item{window}{- the window of data}

\item{weekly}{- either "weekday","weekend" or "none" to define whether to fit a fixed effect model to the weekday, or the is.weekend}

\item{includeModel}{- keep the fitted model as a list column for fit analysis}

\item{...}{}
}
\value{
a dataframe of evaluation dates, growth rates, proportions and model fit
}
\description{
Calculate a slightly more robust estimate of growth rate and proportion based on a single model at a range of
}
