% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{renameResult}
\alias{renameResult}
\title{Rename a growth rate estimate by placing a prefix in front o}
\usage{
renameResult(
  df,
  prefix,
  estimates = c("Growth", "Est", "Proportion", "Rt", "doublingTime")
)
}
\arguments{
\item{df}{the datafram with the Growth rate, possion rate, R_t or proportion estimates}

\item{prefix}{the prefix to add}

\item{estimates}{which estimates to rename (defaults to all of "Growth","Est","Proportion" and "Rt")}
}
\value{
the dataframe with the columns renamed
}
\description{
Rename a growth rate estimate by placing a prefix in front o
}
