% Generated by r6-generator-maven-plugin: do not edit by hand
\docType{package}
\name{roogledocs-package}
\alias{roogledocs}
\alias{roogledocs-package}
\title{roogledocs: R wrapper for Googledocs java library}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}
R wrapper for Googledocs java library

Version: 0.01

Classes:
\itemize{ 
	\item \code{JavaApi}
	\item \code{RoogleDocs}
}
}

\author{
\itemize{ 
   \item Rob Challen rob@terminological.co.uk 0000-0002-5504-7768
   \item terminological   
}
}

\keyword{java api}