% Generated by r6-generator-maven-plugin: do not edit by hand

\name{RoogleDocs}

\alias{RoogleDocs}

\title{RoogleDocs}

\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{tokenDirectory}{tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to "normalizePath('~/.roogledocs', mustWork = FALSE)") - (java expects a String)}
	\if{html}{\out{</div>}}
}


\description{
Programmatically substitute data into a google doc

Version: 0.01

Generated: 2022-04-26T19:20:38.738
}

\details{
	Programmatically substitute data into a google doc
}

\examples{\dontrun{
J = roogledocs::JavaApi$get();
instance = J$RoogleDocs$new(tokenDirectory)

## -----------------------------------
## Method `RoogleDocs$withDocument`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$findOrCreateDocument`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$searchForDocuments`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$tagsDefined`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$updateTaggedText`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$updateTaggedImage`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$revertTags`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$updateTable`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$updateFigure`
## -----------------------------------

}}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$RoogleDocs$new(tokenDirectory)}}
		}
	}
	\subsection{Static methods}{
		\itemize{
			\item{none}
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-withDocument}{\code{instance$withDocument(shareUrlOrDocId)}}
			\item \href{#method-findOrCreateDocument}{\code{instance$findOrCreateDocument(title)}}
			\item \href{#method-searchForDocuments}{\code{instance$searchForDocuments(titleMatch)}}
			\item \href{#method-tagsDefined}{\code{instance$tagsDefined()}}
			\item \href{#method-updateTaggedText}{\code{instance$updateTaggedText(text, tagName)}}
			\item \href{#method-updateTaggedImage}{\code{instance$updateTaggedImage(absoluteFilePath, tagName, dpi)}}
			\item \href{#method-revertTags}{\code{instance$revertTags()}}
			\item \href{#method-updateTable}{\code{instance$updateTable(longFormatTable, tableIndex, colWidths, tableWidthInches)}}
			\item \href{#method-updateFigure}{\code{instance$updateFigure(absoluteFilePath, figureIndex, dpi)}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
J$RoogleDocs$new(tokenDirectory)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to "normalizePath('~/.roogledocs', mustWork = FALSE)")}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 RoogleDocs object: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-withDocument"></a>}}
	
	\subsection{Method \code{withDocument()}}{
		Select a document by its share url or id.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$withDocument(shareUrlOrDocId)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{shareUrlOrDocId the url from clicking a share button in google docs or an id from searchForDocuments() method}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 RoogleDocs object: itself - a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-findOrCreateDocument"></a>}}
	
	\subsection{Method \code{findOrCreateDocument()}}{
		Search for or create a document if it is missing.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$findOrCreateDocument(title)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{title a document title. If there is an exact match in google drive then that document will be used 
otherwise a new one will be created.}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 RoogleDocs object: itself - a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-searchForDocuments"></a>}}
	
	\subsection{Method \code{searchForDocuments()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$searchForDocuments(titleMatch)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{titleMatch a string to be searched for as an approximate match. All results will be retrieved with document ids.}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RDataframe: a dataframe containing id and name columns
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-tagsDefined"></a>}}
	
	\subsection{Method \code{tagsDefined()}}{
		Finds tags defined in the current document
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$tagsDefined()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RDataframe: a dataframe containing tag and count columns
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-updateTaggedText"></a>}}
	
	\subsection{Method \code{updateTaggedText()}}{
		Substitutes all occurrences of {{tag-name}} with the text parameter.
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$updateTaggedText(text, tagName)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{text the value to replace the tag with (e.g. a result from analysis)}{ - (java expects a String)}
				\item{tagName the tag name}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 RoogleDocs object: itself - a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-updateTaggedImage"></a>}}
	
	\subsection{Method \code{updateTaggedImage()}}{
		Substitutes all occurrences of {{tag-name}} with an image from the local storage
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$updateTaggedImage(absoluteFilePath, tagName, dpi)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{absoluteFilePath a file path to an png image file.}{ - (java expects a String)}
				\item{tagName the tag name}{ - (java expects a String)}
				\item{dpi the dots per inch of the image in the document (defaults to 300) - (defaulting to "300")}{ - (java expects a double)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 RoogleDocs object: itself - a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-revertTags"></a>}}
	
	\subsection{Method \code{revertTags()}}{
		remove all text and images inserted by roogledocs and returns the bare document. 
This is needed if content is being moved
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$revertTags()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 RoogleDocs object: itself - a fluent method
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-updateTable"></a>}}
	
	\subsection{Method \code{updateTable()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$updateTable(longFormatTable, tableIndex, colWidths, tableWidthInches)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{longFormatTable A dataframe consisting of the table content and formatting indexed by row and column}{ - (java expects a RDataframe)}
				\item{tableIndex what is the table index in the document? leave out for a new table at the end of the document. - (defaulting to "-1")}{ - (java expects a int)}
				\item{colWidths A vector including the relative length of each column. This can be left out if longFormatTable comes from as.long_format_table - (defaulting to "attr(longFormatTable,'colWidths')")}{ - (java expects a RNumericVector)}
				\item{tableWidthInches The final width of the table in inches (defaults to a size that fits in A4 page with margins) - (defaulting to "6.2")}{ - (java expects a RNumeric)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 RoogleDocs object: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-updateFigure"></a>}}
	
	\subsection{Method \code{updateFigure()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$updateFigure(absoluteFilePath, figureIndex, dpi)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{absoluteFilePath a file path to an png image file.}{ - (java expects a String)}
				\item{figureIndex what is the figure index in the document (only counts inline images - not absolutely positioned ones)? leave out for a new image at the end of the document. - (defaulting to "-1")}{ - (java expects a int)}
				\item{dpi the dots per inch of the image in the document (defaults to 300) - (defaulting to "300")}{ - (java expects a double)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 RoogleDocs object: itself (a fluent method)
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
}