% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{rtFromGrowthRate}
\alias{rtFromGrowthRate}
\title{Calculate a reproduction number estimate using the Wallinga 2007 estimation using empirical generation time distribution. This uses resampling to transmit uncertainty in growth rate estimates}
\usage{
rtFromGrowthRate(
  simpleTimeseries,
  infectivityProfile,
  yMatrix = infectivityProfile$yMatrix,
  aVector = infectivityProfile$aVector,
  bootstraps = 20 * dim(yMatrix)[2],
  quantiles = c(0.025, 0.05, 0.25, 0.5, 0.75, 0.95, 0.975)
)
}
\arguments{
\item{simpleTimeseries}{- With a "Growth" estimate as a normally distributed quantility}

\item{yMatrix}{- the matrix of possible infectivity profiles as discrete distributions}

\item{aVector}{- the upper boundaries of the time cut-offs for the infectivity profiles}

\item{bootstraps}{- the number of bootstraps to take to calculate for each point.}

\item{quantiles}{- quantiles to calculate.}
}
\value{
a timeseries with "Rt" estimates
}
\description{
Calculate a reproduction number estimate using the Wallinga 2007 estimation using empirical generation time distribution. This uses resampling to transmit uncertainty in growth rate estimates
}
