% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{Serialiser}

\alias{Serialiser}

\title{Serialiser}

\usage{\preformatted{
library(testRapi)
J = JavaApi$new();
instance <- J$Serialiser$new(...);
}}

\description{
	Actually useful class which takes an R dataframe and a filename
}

\details{
	
string and serialises the dataframe so it can be used natively in java
for testing purposes.
}

\examples{
## -----------------------------------
## Method `Serialiser$serialiseDataframe`
## -----------------------------------

## -----------------------------------
## Method `Serialiser$serialiseList`
## -----------------------------------

## -----------------------------------
## Method `Serialiser$serialiseNamedList`
## -----------------------------------

}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$Serialiser$new()}}
		}
	}
	\subsection{Static methods}{
		\itemize{
			\item \href{#method-serialiseDataframe}{\code{J$Serialiser$serialiseDataframe(dataframe, filename)}}
			\item \href{#method-serialiseList}{\code{J$Serialiser$serialiseList(dataframe, filename)}}
			\item \href{#method-serialiseNamedList}{\code{J$Serialiser$serialiseNamedList(dataframe, filename)}}
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$Serialiser$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 Serialiser object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-serialiseDataframe"></a>}}
	
	\subsection{Method \code{serialiseDataframe()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$Serialiser$serialiseDataframe(dataframe, filename)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{dataframe}{ - (java expects a RDataframe)}
				\item{filename}{ - (java expects a String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-serialiseList"></a>}}
	
	\subsection{Method \code{serialiseList()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$Serialiser$serialiseList(dataframe, filename)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{dataframe}{ - (java expects a RList)}
				\item{filename}{ - (java expects a String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-serialiseNamedList"></a>}}
	
	\subsection{Method \code{serialiseNamedList()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$Serialiser$serialiseNamedList(dataframe, filename)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{dataframe}{ - (java expects a RNamedList)}
				\item{filename}{ - (java expects a String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}