% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic-datasets.R
\name{uncertainGammaInfectivityProfiles}
\alias{uncertainGammaInfectivityProfiles}
\title{Generate a set of infectivity profiles based on uncertain}
\usage{
uncertainGammaInfectivityProfiles(
  meanOfMean,
  sdOfMean,
  meanOfSd,
  sdOfSd,
  n = 100,
  range = 20
)
}
\arguments{
\item{meanOfMean}{- the mean value of the mean of the gamma infectivity profile}

\item{sdOfMean}{- the sd value of the mean of the gamma infectivity profile}

\item{meanOfSd}{- the mean value of the sd of the gamma infectivity profile}

\item{sdOfSd}{- the sd value of the sd of the gamma infectivity profile}

\item{n}{- number of samples}

\item{range}{- length of desired infectivity profile output}
}
\value{
a set of infectivity profiles
}
\description{
Generate a set of infectivity profiles based on uncertain
}
