% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{weekendEffect}
\alias{weekendEffect}
\title{Calculates a weighting to apply to each day of week}
\usage{
weekendEffect(simpleTimeseries, valueVar = "value", ...)
}
\arguments{
\item{simpleTimeseries}{a covid timeseries data frame}

\item{valueVar}{the variable with the weekly periodicity}

\item{...}{}
}
\value{
the dataframe with a weekday.wt column which says how much that value is over expressed in the data
}
\description{
Calculates a weighting to apply to each day of week
}
