% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{BounceTest}

\alias{BounceTest}

\title{BounceTest}

\usage{\preformatted{
library(testRapi)
J = JavaApi$new();
instance <- J$BounceTest$new(...);
}}

\description{
	
}

\details{
	
}

\examples{
## -----------------------------------
## Method `BounceTest$bounceVoid`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceInt`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceDouble`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceString`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceNull`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceCharacter`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceCharacterVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceInteger`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceIntegerVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceDate`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceDateVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceFactor`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceFactorVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceLogical`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceLogicalVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceNumeric`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceNumericVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceDataframe`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceList`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceNamedList`
## -----------------------------------

}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$BounceTest$new()}}
		}
	}
	\subsection{Static methods}{
		\itemize{
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-bounceVoid}{\code{instance$bounceVoid()}}
			\item \href{#method-bounceInt}{\code{instance$bounceInt(x)}}
			\item \href{#method-bounceDouble}{\code{instance$bounceDouble(x)}}
			\item \href{#method-bounceString}{\code{instance$bounceString(x)}}
			\item \href{#method-bounceNull}{\code{instance$bounceNull(x)}}
			\item \href{#method-bounceCharacter}{\code{instance$bounceCharacter(x)}}
			\item \href{#method-bounceCharacterVector}{\code{instance$bounceCharacterVector(x)}}
			\item \href{#method-bounceInteger}{\code{instance$bounceInteger(x)}}
			\item \href{#method-bounceIntegerVector}{\code{instance$bounceIntegerVector(x)}}
			\item \href{#method-bounceDate}{\code{instance$bounceDate(x)}}
			\item \href{#method-bounceDateVector}{\code{instance$bounceDateVector(x)}}
			\item \href{#method-bounceFactor}{\code{instance$bounceFactor(x)}}
			\item \href{#method-bounceFactorVector}{\code{instance$bounceFactorVector(x)}}
			\item \href{#method-bounceLogical}{\code{instance$bounceLogical(x)}}
			\item \href{#method-bounceLogicalVector}{\code{instance$bounceLogicalVector(x)}}
			\item \href{#method-bounceNumeric}{\code{instance$bounceNumeric(x)}}
			\item \href{#method-bounceNumericVector}{\code{instance$bounceNumericVector(x)}}
			\item \href{#method-bounceDataframe}{\code{instance$bounceDataframe(x)}}
			\item \href{#method-bounceList}{\code{instance$bounceList(x)}}
			\item \href{#method-bounceNamedList}{\code{instance$bounceNamedList(x)}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$BounceTest$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 BounceTest object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceVoid"></a>}}
	
	\subsection{Method \code{bounceVoid()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceVoid()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@221a2068: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceInt"></a>}}
	
	\subsection{Method \code{bounceInt()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceInt(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@5a49af50)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@5a49af50: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDouble"></a>}}
	
	\subsection{Method \code{bounceDouble()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceDouble(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@4f8d86e4)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@4f8d86e4: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceString"></a>}}
	
	\subsection{Method \code{bounceString()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceString(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@60dd0587)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@60dd0587: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNull"></a>}}
	
	\subsection{Method \code{bounceNull()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceNull(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@5f631ca0)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@5f631ca0: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceCharacter"></a>}}
	
	\subsection{Method \code{bounceCharacter()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceCharacter(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@7930ffa9)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@7930ffa9: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceCharacterVector"></a>}}
	
	\subsection{Method \code{bounceCharacterVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceCharacterVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@1697f2b3)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@1697f2b3: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceInteger"></a>}}
	
	\subsection{Method \code{bounceInteger()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceInteger(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@12ad1b2a)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@12ad1b2a: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceIntegerVector"></a>}}
	
	\subsection{Method \code{bounceIntegerVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceIntegerVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@3134153d)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@3134153d: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDate"></a>}}
	
	\subsection{Method \code{bounceDate()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceDate(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@767599a7)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@767599a7: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDateVector"></a>}}
	
	\subsection{Method \code{bounceDateVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceDateVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@5f5effb0)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@5f5effb0: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceFactor"></a>}}
	
	\subsection{Method \code{bounceFactor()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceFactor(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@25d0cb3a)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@25d0cb3a: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceFactorVector"></a>}}
	
	\subsection{Method \code{bounceFactorVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceFactorVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@345cbf40)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@345cbf40: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceLogical"></a>}}
	
	\subsection{Method \code{bounceLogical()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceLogical(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@6e3eb0cd)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@6e3eb0cd: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceLogicalVector"></a>}}
	
	\subsection{Method \code{bounceLogicalVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceLogicalVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@463561c5)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@463561c5: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNumeric"></a>}}
	
	\subsection{Method \code{bounceNumeric()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceNumeric(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@dd2856e)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@dd2856e: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNumericVector"></a>}}
	
	\subsection{Method \code{bounceNumericVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceNumericVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@659feb22)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@659feb22: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDataframe"></a>}}
	
	\subsection{Method \code{bounceDataframe()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceDataframe(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@77f905e3)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@77f905e3: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceList"></a>}}
	
	\subsection{Method \code{bounceList()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceList(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@3468ee6e)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@3468ee6e: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNamedList"></a>}}
	
	\subsection{Method \code{bounceNamedList()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceNamedList(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a uk.co.terminological.jsr233plugin.RType@2f4b98f6)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@2f4b98f6: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}