% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{BounceTest}

\alias{BounceTest}

\title{BounceTest}

\usage{\preformatted{
library(testRapi)
J = JavaApi$new();
instance <- J$BounceTest$new(...);
}}

\description{
	
}

\details{
	
}

\examples{
## -----------------------------------
## Method `BounceTest$bounceVoid`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceInt`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceDouble`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceString`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceNull`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceCharacter`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceCharacterVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceInteger`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceIntegerVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceDate`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceDateVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceFactor`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceFactorVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceLogical`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceLogicalVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceNumeric`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceNumericVector`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceDataframe`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceList`
## -----------------------------------

## -----------------------------------
## Method `BounceTest$bounceNamedList`
## -----------------------------------

}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$BounceTest$new()}}
		}
	}
	\subsection{Static methods}{
		\itemize{
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-bounceVoid}{\code{instance$bounceVoid()}}
			\item \href{#method-bounceInt}{\code{instance$bounceInt(x)}}
			\item \href{#method-bounceDouble}{\code{instance$bounceDouble(x)}}
			\item \href{#method-bounceString}{\code{instance$bounceString(x)}}
			\item \href{#method-bounceNull}{\code{instance$bounceNull(x)}}
			\item \href{#method-bounceCharacter}{\code{instance$bounceCharacter(x)}}
			\item \href{#method-bounceCharacterVector}{\code{instance$bounceCharacterVector(x)}}
			\item \href{#method-bounceInteger}{\code{instance$bounceInteger(x)}}
			\item \href{#method-bounceIntegerVector}{\code{instance$bounceIntegerVector(x)}}
			\item \href{#method-bounceDate}{\code{instance$bounceDate(x)}}
			\item \href{#method-bounceDateVector}{\code{instance$bounceDateVector(x)}}
			\item \href{#method-bounceFactor}{\code{instance$bounceFactor(x)}}
			\item \href{#method-bounceFactorVector}{\code{instance$bounceFactorVector(x)}}
			\item \href{#method-bounceLogical}{\code{instance$bounceLogical(x)}}
			\item \href{#method-bounceLogicalVector}{\code{instance$bounceLogicalVector(x)}}
			\item \href{#method-bounceNumeric}{\code{instance$bounceNumeric(x)}}
			\item \href{#method-bounceNumericVector}{\code{instance$bounceNumericVector(x)}}
			\item \href{#method-bounceDataframe}{\code{instance$bounceDataframe(x)}}
			\item \href{#method-bounceList}{\code{instance$bounceList(x)}}
			\item \href{#method-bounceNamedList}{\code{instance$bounceNamedList(x)}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$BounceTest$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 BounceTest object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceVoid"></a>}}
	
	\subsection{Method \code{bounceVoid()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceVoid()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceInt"></a>}}
	
	\subsection{Method \code{bounceInt()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceInt(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a int)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		int: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDouble"></a>}}
	
	\subsection{Method \code{bounceDouble()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceDouble(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a double)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		double: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceString"></a>}}
	
	\subsection{Method \code{bounceString()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceString(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		String: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNull"></a>}}
	
	\subsection{Method \code{bounceNull()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceNull(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RNull)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RNull: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceCharacter"></a>}}
	
	\subsection{Method \code{bounceCharacter()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceCharacter(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RCharacter)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RCharacter: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceCharacterVector"></a>}}
	
	\subsection{Method \code{bounceCharacterVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceCharacterVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RCharacterVector)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RCharacterVector: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceInteger"></a>}}
	
	\subsection{Method \code{bounceInteger()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceInteger(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RInteger)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RInteger: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceIntegerVector"></a>}}
	
	\subsection{Method \code{bounceIntegerVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceIntegerVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RIntegerVector)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RIntegerVector: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDate"></a>}}
	
	\subsection{Method \code{bounceDate()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceDate(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RDate)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RDate: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDateVector"></a>}}
	
	\subsection{Method \code{bounceDateVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceDateVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RDateVector)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RDateVector: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceFactor"></a>}}
	
	\subsection{Method \code{bounceFactor()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceFactor(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RFactor)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RFactor: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceFactorVector"></a>}}
	
	\subsection{Method \code{bounceFactorVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceFactorVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RFactorVector)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RFactorVector: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceLogical"></a>}}
	
	\subsection{Method \code{bounceLogical()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceLogical(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RLogical)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RLogical: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceLogicalVector"></a>}}
	
	\subsection{Method \code{bounceLogicalVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceLogicalVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RLogicalVector)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RLogicalVector: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNumeric"></a>}}
	
	\subsection{Method \code{bounceNumeric()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceNumeric(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RNumeric)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RNumeric: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNumericVector"></a>}}
	
	\subsection{Method \code{bounceNumericVector()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceNumericVector(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RNumericVector)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RNumericVector: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceDataframe"></a>}}
	
	\subsection{Method \code{bounceDataframe()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceDataframe(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RDataframe)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RDataframe: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceList"></a>}}
	
	\subsection{Method \code{bounceList()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceList(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RList)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RList: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-bounceNamedList"></a>}}
	
	\subsection{Method \code{bounceNamedList()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$BounceTest$new(...);
  instance$bounceNamedList(x)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{x}{ - (java expects a RNamedList)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RNamedList: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}