% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{FactoryTest}

\alias{FactoryTest}

\title{FactoryTest}



\description{
missing description

Version: 0.01

Generated: 2020-11-03T21:33:22.947
}

\details{
	no details
}

\examples{\dontrun{
J = testRapi::JavaApi$get();
instance = J$FactoryTest$new()

## -----------------------------------
## Method `FactoryTest$generateCharacter`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateNumeric`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateInteger`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateFactor`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateLogical`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateCharacterVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateNumericVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateIntegerVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateFactorVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateLogicalVec`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateDataframe`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateStreamDataframe`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateList`
## -----------------------------------

## -----------------------------------
## Method `FactoryTest$generateNamedList`
## -----------------------------------

}}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$FactoryTest$new()}}
		}
	}
	\subsection{Static methods}{
		\itemize{
			\item{none}
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-generateCharacter}{\code{instance$generateCharacter()}}
			\item \href{#method-generateNumeric}{\code{instance$generateNumeric()}}
			\item \href{#method-generateInteger}{\code{instance$generateInteger()}}
			\item \href{#method-generateFactor}{\code{instance$generateFactor()}}
			\item \href{#method-generateLogical}{\code{instance$generateLogical()}}
			\item \href{#method-generateCharacterVec}{\code{instance$generateCharacterVec()}}
			\item \href{#method-generateNumericVec}{\code{instance$generateNumericVec()}}
			\item \href{#method-generateIntegerVec}{\code{instance$generateIntegerVec()}}
			\item \href{#method-generateFactorVec}{\code{instance$generateFactorVec()}}
			\item \href{#method-generateLogicalVec}{\code{instance$generateLogicalVec()}}
			\item \href{#method-generateDataframe}{\code{instance$generateDataframe()}}
			\item \href{#method-generateStreamDataframe}{\code{instance$generateStreamDataframe()}}
			\item \href{#method-generateList}{\code{instance$generateList()}}
			\item \href{#method-generateNamedList}{\code{instance$generateNamedList()}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		the default no-args constructor
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$FactoryTest$new()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 FactoryTest object: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateCharacter"></a>}}
	
	\subsection{Method \code{generateCharacter()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateCharacter()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RCharacter: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateNumeric"></a>}}
	
	\subsection{Method \code{generateNumeric()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateNumeric()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RNumeric: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateInteger"></a>}}
	
	\subsection{Method \code{generateInteger()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateInteger()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RInteger: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateFactor"></a>}}
	
	\subsection{Method \code{generateFactor()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateFactor()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RFactor: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateLogical"></a>}}
	
	\subsection{Method \code{generateLogical()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateLogical()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RLogical: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateCharacterVec"></a>}}
	
	\subsection{Method \code{generateCharacterVec()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateCharacterVec()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RCharacterVector: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateNumericVec"></a>}}
	
	\subsection{Method \code{generateNumericVec()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateNumericVec()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RNumericVector: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateIntegerVec"></a>}}
	
	\subsection{Method \code{generateIntegerVec()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateIntegerVec()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RIntegerVector: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateFactorVec"></a>}}
	
	\subsection{Method \code{generateFactorVec()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateFactorVec()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RFactorVector: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateLogicalVec"></a>}}
	
	\subsection{Method \code{generateLogicalVec()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateLogicalVec()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RLogicalVector: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateDataframe"></a>}}
	
	\subsection{Method \code{generateDataframe()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateDataframe()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RDataframe: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateStreamDataframe"></a>}}
	
	\subsection{Method \code{generateStreamDataframe()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateStreamDataframe()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RDataframe: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateList"></a>}}
	
	\subsection{Method \code{generateList()}}{
		Lists are much harder to type check than vectors hence RList builder methods throw checked exceptions
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateList()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RList: a RList containing the supplied Java objects converted into RObjects
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateNamedList"></a>}}
	
	\subsection{Method \code{generateNamedList()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$FactoryTest$new(...);
instance$generateNamedList()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RNamedList: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
}