% Generated by r6-generator-maven-plugin: do not edit by hand
\docType{package}
\name{html2pdfr-package}
\alias{html2pdfr}
\alias{html2pdfr-package}
\title{html2pdfr: R wrapper for OpenHTMLtoPDF java library}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}
R wrapper for OpenHTMLtoPDF java library.

Version: 0.01

Classes:
\itemize{ 
	\item \code{JavaApi}
	\item \code{HtmlConverter}
}
}

\author{
\itemize{ 
   \item Rob Challen rob.challen@bristol.ac.uk 0000-0002-5504-7768
}
}

\keyword{java api}