% Generated by r6-generator-maven-plugin: do not edit by hand

\name{JavaApi}

\alias{JavaApi}

\title{R wrapper for OpenHTMLtoPDF java library}

\section{Usage}{
	\if{html}{\out{<div class="r">}}
	\preformatted{
 J = html2pdfr::JavaApi$get(logLevel)
 	}
  \if{html}{\out{</div>}}
}

\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{logLevel}{optional - the slf4j log level as a string - one of OFF (most specific, no logging), 
	FATAL (most specific, little data), ERROR, WARN, INFO, DEBUG, 
	TRACE (least specific, a lot of data), ALL (least specific, all data)}
	\if{html}{\out{</div>}}
}

\description{
R wrapper for OpenHTMLtoPDF java library.

Version: 0.01

Generated: 2022-03-10T16:36:40.964
}

\author{\email{rob.challen@bristol.ac.uk}}

\keyword{java api}

\section{Static methods and constructors}{
	\itemize{
		\item \code{JavaApi$get()}
		\item \href{#method-HtmlConverter-new}{\code{J$HtmlConverter$new(fontfiles)}}
	}
	

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-HtmlConverter-new"></a>}}
	
	\subsection{Method \code{HtmlConverter$new()}}{
		the default no-args constructor
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = html2pdfr::JavaApi$get()
J$HtmlConverter$new(fontfiles)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{fontfiles}{ - (java expects a RCharacterVector)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 HtmlConverter object: 
		}
	}

}