% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{JavaApi}

\alias{JavaApi}

\title{A test library}

\usage{\preformatted{
library(testRapi)
J = JavaApi$new();
}}

\description{
	An optional long description of the package
}

\keyword{java api}

\section{Methods}{
	\subsection{Public methods}{
		\itemize{
			\item \code{JavaApi$new()}
	\item \href{#method-MoreFeatureTest-new}{\code{J$MoreFeatureTest$new(message1, message2)}}
			\item \href{#method-MoreFeatureTest-create}{\code{J$MoreFeatureTest$create(message1, message2)}}
			\item \href{#method-MoreFeatureTest-concat}{\code{J$MoreFeatureTest$concat(message1, message2)}}
	\item \href{#method-BounceTest-new}{\code{J$BounceTest$new()}}
	\item \href{#method-FeatureTest-new}{\code{J$FeatureTest$new(logMessage)}}
			\item \href{#method-FeatureTest-demoStatic}{\code{J$FeatureTest$demoStatic(message)}}
			\item \href{#method-FeatureTest-diamonds}{\code{J$FeatureTest$diamonds()}}
	\item \href{#method-FactoryTest-new}{\code{J$FactoryTest$new()}}
	\item \href{#method-Serialiser-new}{\code{J$Serialiser$new()}}
			\item \href{#method-Serialiser-serialiseDataframe}{\code{J$Serialiser$serialiseDataframe(dataframe, filename)}}
			\item \href{#method-Serialiser-serialiseList}{\code{J$Serialiser$serialiseList(dataframe, filename)}}
			\item \href{#method-Serialiser-serialiseNamedList}{\code{J$Serialiser$serialiseNamedList(dataframe, filename)}}
	\item \href{#method-MinimalExample-new}{\code{J$MinimalExample$new()}}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{MoreFeatureTest$new()}}{
		the first constructor is used if there are none annotated
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$MoreFeatureTest$new(message1, message2)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message1 - the message to be printed}{ - (java expects a RCharacter)}
				\item{message2 - will be used for toString}{ - (java expects a RCharacter)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 MoreFeatureTest object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-create"></a>}}
	
	\subsection{Method \code{MoreFeatureTest$create()}}{
		A static object constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$MoreFeatureTest$create(message1, message2)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message1}{ - (java expects a RCharacter)}
				\item{message2}{ - (java expects a RCharacter)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 MoreFeatureTest object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
  J = testRapi::JavaApi$new()
  J$MoreFeatureTest$create('Hello,',' World')
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-concat"></a>}}
	
	\subsection{Method \code{MoreFeatureTest$concat()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$MoreFeatureTest$concat(message1, message2)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message1}{ - (java expects a RCharacter)}
				\item{message2}{ - (java expects a RCharacter)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RCharacter: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{BounceTest$new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$BounceTest$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 BounceTest object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{FeatureTest$new()}}{
		A maximum of one constructor of any signature can be used. </br>

If different constructors are needed then they may be used but not 
included in the R Api (i.e. not annotated with @RMethod. </br>

Static factory methods can be used instead.
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$FeatureTest$new(logMessage)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{logMessage - a message which will be logged}{ - (java expects a String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 FeatureTest object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
    
			J = testRapi::JavaApi$new()
  minExample = J$FeatureTest$new('Hello, java constructor!')
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-demoStatic"></a>}}
	
	\subsection{Method \code{FeatureTest$demoStatic()}}{
		Static methods are also supported. These are accessed through the
root of the R api.
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$FeatureTest$demoStatic(message)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message}{ - (java expects a String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
  J = testRapi::JavaApi$new()
  J$FeatureTest$demoStatic('Ola, el mundo')
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-diamonds"></a>}}
	
	\subsection{Method \code{FeatureTest$diamonds()}}{
		A copy of the ggplot2::diamonds dataframe serialised into java, using
RObject.writeRDS, saved within the jar file of the package, and exposed here
using RObject.readRDS.
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$FeatureTest$diamonds()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		RDataframe: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
  J = testRapi::JavaApi$new()
  J$FeatureTest$diamonds()
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{FactoryTest$new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$FactoryTest$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 FactoryTest object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{Serialiser$new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$Serialiser$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 Serialiser object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-serialiseDataframe"></a>}}
	
	\subsection{Method \code{Serialiser$serialiseDataframe()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$Serialiser$serialiseDataframe(dataframe, filename)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{dataframe}{ - (java expects a RDataframe)}
				\item{filename}{ - (java expects a String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-serialiseList"></a>}}
	
	\subsection{Method \code{Serialiser$serialiseList()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$Serialiser$serialiseList(dataframe, filename)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{dataframe}{ - (java expects a RList)}
				\item{filename}{ - (java expects a String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-serialiseNamedList"></a>}}
	
	\subsection{Method \code{Serialiser$serialiseNamedList()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$Serialiser$serialiseNamedList(dataframe, filename)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{dataframe}{ - (java expects a RNamedList)}
				\item{filename}{ - (java expects a String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{MinimalExample$new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$MinimalExample$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 MinimalExample object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}