% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{JavaApi}

\alias{JavaApi}

\title{A test library}

\usage{\preformatted{
library(testRapi)
J = JavaApi$new();
}}

\description{
	An optional long description of the package
}

\keyword{java api}

\section{Methods}{
	\subsection{Public methods}{
		\itemize{
			\item \code{JavaApi$new()}
	\item \href{#method-TestRApi-new}{\code{J$TestRApi$new()}}
	\item \href{#method-BounceTest-new}{\code{J$BounceTest$new()}}
	\item \href{#method-FactoryTest-new}{\code{J$FactoryTest$new()}}
	\item \href{#method-AnotherTestRApi-new}{\code{J$AnotherTestRApi$new(message1, message2)}}
			\item \href{#method-AnotherTestRApi-create}{\code{J$AnotherTestRApi$create(message1, message2)}}
			\item \href{#method-AnotherTestRApi-concat}{\code{J$AnotherTestRApi$concat(message1, message2)}}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{TestRApi$new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$TestRApi$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 TestRApi object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{BounceTest$new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$BounceTest$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 BounceTest object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{FactoryTest$new()}}{
		the default no-args constructor
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$FactoryTest$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 FactoryTest object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{AnotherTestRApi$new()}}{
		the first constructor is used if there are none annotated
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$AnotherTestRApi$new(message1, message2)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message1 - the message to be printed}{ - (java expects a uk.co.terminological.jsr233plugin.RType@7930ffa9)}
				\item{message2 - will be used for toString}{ - (java expects a uk.co.terminological.jsr233plugin.RType@7930ffa9)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 AnotherTestRApi object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-create"></a>}}
	
	\subsection{Method \code{AnotherTestRApi$create()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$AnotherTestRApi$create(message1, message2)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message1}{ - (java expects a uk.co.terminological.jsr233plugin.RType@7930ffa9)}
				\item{message2}{ - (java expects a uk.co.terminological.jsr233plugin.RType@7930ffa9)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 AnotherTestRApi object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-concat"></a>}}
	
	\subsection{Method \code{AnotherTestRApi$concat()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$AnotherTestRApi$concat(message1, message2)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message1}{ - (java expects a uk.co.terminological.jsr233plugin.RType@7930ffa9)}
				\item{message2}{ - (java expects a uk.co.terminological.jsr233plugin.RType@7930ffa9)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		uk.co.terminological.jsr233plugin.RType@7930ffa9: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}