% Generated by r6-generator-maven-plugin: do not edit by hand

\name{JavaApi}

\alias{JavaApi}

\title{A test library}

\section{Usage}{
	\if{html}{\out{<div class="r">}}
	\preformatted{
 J = testRapi::JavaApi$get(logLevel)
 	}
  \if{html}{\out{</div>}}
}

\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{logLevel}{optional - the slf4j log level as a string - one of OFF (most specific, no logging), 
	FATAL (most specific, little data), ERROR, WARN, INFO, DEBUG, 
	TRACE (least specific, a lot of data), ALL (least specific, all data)}
	\if{html}{\out{</div>}}
}

\description{
This package is a feature test of the R6 Generator Maven Plugin. It is not intended to be useful beyond testing, demonstrating and documenting the features of the plugin.

Version: 0.2.0.9000

Generated: 2022-05-23T16:12:12.726871
}

\author{\email{rc538@exeter.ac.uk}}

\keyword{java api}

\section{Static methods and constructors}{
	\itemize{
		\item \code{JavaApi$get()}
		\item \href{#method-MoreFeatureTest-new}{\code{J$MoreFeatureTest$new(message1, message2)}}
		\item \href{#method-MoreFeatureTest-create}{\code{J$MoreFeatureTest$create(message1, message2)}}
		\item \href{#method-MoreFeatureTest-concat}{\code{J$MoreFeatureTest$concat(message1, message2)}}
		\item \href{#method-BounceTest-new}{\code{J$BounceTest$new()}}
		\item \href{#method-FeatureTest-new}{\code{J$FeatureTest$new(logMessage)}}
		\item \href{#method-FeatureTest-demoStatic}{\code{J$FeatureTest$demoStatic(message)}}
		\item \href{#method-FeatureTest-diamonds}{\code{J$FeatureTest$diamonds()}}
		\item \href{#method-FactoryTest-new}{\code{J$FactoryTest$new()}}
		\item \href{#method-Serialiser-new}{\code{J$Serialiser$new()}}
		\item \href{#method-Serialiser-serialiseDataframe}{\code{J$Serialiser$serialiseDataframe(dataframe, filename)}}
		\item \href{#method-Serialiser-deserialiseDataframe}{\code{J$Serialiser$deserialiseDataframe(filename)}}
		\item \href{#method-Serialiser-serialiseList}{\code{J$Serialiser$serialiseList(dataframe, filename)}}
		\item \href{#method-Serialiser-deserialiseList}{\code{J$Serialiser$deserialiseList(filename)}}
		\item \href{#method-Serialiser-serialiseNamedList}{\code{J$Serialiser$serialiseNamedList(dataframe, filename)}}
		\item \href{#method-Serialiser-deserialiseNamedList}{\code{J$Serialiser$deserialiseNamedList(filename)}}
		\item \href{#method-MinimalExample-new}{\code{J$MinimalExample$new()}}
	}
	

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-MoreFeatureTest-new"></a>}}
	
	\subsection{Method \code{MoreFeatureTest$new()}}{
		the first constructor is used if there are none annotated
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$MoreFeatureTest$new(message1, message2)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{message1 - the message to be printed}{ - (java expects a RCharacter)}
					\item{message2 - will be used for toString}{ - (java expects a RCharacter)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 MoreFeatureTest object: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-MoreFeatureTest-create"></a>}}
	
	\subsection{Method \code{MoreFeatureTest$create()}}{
		A static object constructor
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$MoreFeatureTest$create(message1, message2)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{message1 - the message to be printed}{ - (java expects a RCharacter)}
					\item{message2 - will be used for toString}{ - (java expects a RCharacter)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 MoreFeatureTest object: A MoreFeatureTest R6 object
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-MoreFeatureTest-concat"></a>}}
	
	\subsection{Method \code{MoreFeatureTest$concat()}}{
		no description
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$MoreFeatureTest$concat(message1, message2)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{message1}{ - (java expects a RCharacter)}
					\item{message2}{ - (java expects a RCharacter)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RCharacter: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-BounceTest-new"></a>}}
	
	\subsection{Method \code{BounceTest$new()}}{
		the default no-args constructor
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$BounceTest$new()
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{none}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 BounceTest object: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-FeatureTest-new"></a>}}
	
	\subsection{Method \code{FeatureTest$new()}}{
		A maximum of one constructor of any signature can be used. <br>

If different constructors are needed then they may be used but not 
included in the R Api (i.e. not annotated with @RMethod.) <br>

Static factory methods can be used instead.
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$FeatureTest$new(logMessage)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{logMessage - a message which will be logged}{ - (java expects a String)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 FeatureTest object: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-FeatureTest-demoStatic"></a>}}
	
	\subsection{Method \code{FeatureTest$demoStatic()}}{
		Static methods are also supported. These are accessed through the
root of the R api.
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$FeatureTest$demoStatic(message)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{message a message}{ - (java expects a String)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			void: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-FeatureTest-diamonds"></a>}}
	
	\subsection{Method \code{FeatureTest$diamonds()}}{
		A copy of the ggplot2::diamonds dataframe serialised into java, using
RObject.writeRDS, saved within the jar file of the package, and exposed here
using RObject.readRDS.
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$FeatureTest$diamonds()
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{none}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RDataframe: the ggplot2::diamonds dataframe
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-FactoryTest-new"></a>}}
	
	\subsection{Method \code{FactoryTest$new()}}{
		the default no-args constructor
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$FactoryTest$new()
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{none}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 FactoryTest object: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-Serialiser-new"></a>}}
	
	\subsection{Method \code{Serialiser$new()}}{
		the default no-args constructor
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$Serialiser$new()
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{none}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 Serialiser object: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-Serialiser-serialiseDataframe"></a>}}
	
	\subsection{Method \code{Serialiser$serialiseDataframe()}}{
		no description
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$Serialiser$serialiseDataframe(dataframe, filename)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{dataframe}{ - (java expects a RDataframe)}
					\item{filename}{ - (java expects a String)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			void: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-Serialiser-deserialiseDataframe"></a>}}
	
	\subsection{Method \code{Serialiser$deserialiseDataframe()}}{
		no description
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$Serialiser$deserialiseDataframe(filename)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{filename}{ - (java expects a String)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RDataframe: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-Serialiser-serialiseList"></a>}}
	
	\subsection{Method \code{Serialiser$serialiseList()}}{
		no description
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$Serialiser$serialiseList(dataframe, filename)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{dataframe}{ - (java expects a RList)}
					\item{filename}{ - (java expects a String)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			void: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-Serialiser-deserialiseList"></a>}}
	
	\subsection{Method \code{Serialiser$deserialiseList()}}{
		no description
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$Serialiser$deserialiseList(filename)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{filename}{ - (java expects a String)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RList: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-Serialiser-serialiseNamedList"></a>}}
	
	\subsection{Method \code{Serialiser$serialiseNamedList()}}{
		no description
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$Serialiser$serialiseNamedList(dataframe, filename)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{dataframe}{ - (java expects a RNamedList)}
					\item{filename}{ - (java expects a String)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			void: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-Serialiser-deserialiseNamedList"></a>}}
	
	\subsection{Method \code{Serialiser$deserialiseNamedList()}}{
		no description
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$Serialiser$deserialiseNamedList(filename)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{filename}{ - (java expects a String)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RNamedList: 
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-MinimalExample-new"></a>}}
	
	\subsection{Method \code{MinimalExample$new()}}{
		the default no-args constructor
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$MinimalExample$new()
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{none}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 MinimalExample object: 
		}
	}

}