% Generated by r6-generator-maven-plugin: do not edit by hand

\name{JavaApi}

\alias{JavaApi}

\title{R wrapper for Googledocs java library}

\section{Usage}{
	\if{html}{\out{<div class="r">}}
	\preformatted{
 J = roogledocs::JavaApi$get(logLevel)
 	}
  \if{html}{\out{</div>}}
}

\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{logLevel}{optional - the slf4j log level as a string - one of OFF (most specific, no logging), 
	FATAL (most specific, little data), ERROR, WARN, INFO, DEBUG, 
	TRACE (least specific, a lot of data), ALL (least specific, all data)}
	\if{html}{\out{</div>}}
}

\description{
R wrapper for Googledocs java library

Version: 0.01

Generated: 2022-04-25T18:01:03.169
}

\author{\email{rob@terminological.co.uk}}

\keyword{java api}

\section{Static methods and constructors}{
	\itemize{
		\item \code{JavaApi$get()}
		\item \href{#method-RoogleDocs-new}{\code{J$RoogleDocs$new(tokenDirectory)}}
	}
	

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-RoogleDocs-new"></a>}}
	
	\subsection{Method \code{RoogleDocs$new()}}{
		no description
	
		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = roogledocs::JavaApi$get()
J$RoogleDocs$new(tokenDirectory)
	  		}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\describe{
				\itemize{
					\item{tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to "normalizePath('~/.roogledocs', mustWork = FALSE)")}{ - (java expects a String)}
				}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 RoogleDocs object: 
		}
	}

}