% Generated by r6-generator-maven-plugin: do not edit by hand

\name{MinimalExample}

\alias{MinimalExample}

\title{MinimalExample}



\description{
This class is a very basic example of the features of the rJava maven plugin. <br>

Version: 0.2.0.9000

Generated: 2022-05-23T15:53:14.177619
}

\details{
	
The class is annotated with an @RClass to identify it as part of the R API. <br>
}

\examples{\dontrun{
J = testRapi::JavaApi$get();
instance = J$MinimalExample$new()

## -----------------------------------
## Method `MinimalExample$demo`
## -----------------------------------
  
		minExample = J$MinimalExample$new()
minExample$demo(dataframe=tibble::tibble(input=c(1,2,3)), message='Hello world')

}}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$MinimalExample$new()}}
		}
	}
	\subsection{Static methods}{
		\itemize{
			\item{none}
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-demo}{\code{instance$demo(dataframe, message)}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		the default no-args constructor
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$MinimalExample$new()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 MinimalExample object: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-demo"></a>}}
	
	\subsection{Method \code{demo()}}{
		Documentation of the method can be done in JavaDoc and these will be present in the R documentation
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$MinimalExample$new(...);
instance$demo(dataframe, message)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{dataframe - a dataframe with an arbitrary number of columns}{ - (java expects a RDataframe)}
				\item{message - a message}{ - (java expects a String)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RDataframe: the dataframe unchanged
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
  
		minExample = J$MinimalExample$new()
minExample$demo(dataframe=tibble::tibble(input=c(1,2,3)), message='Hello world')
			}
			\if{html}{\out{</div>}}
		}
	}
}