% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{MoreFeatureTest}

\alias{MoreFeatureTest}

\title{MoreFeatureTest}

\arguments{
	\if{html}{\out{<div class="arguments">}}
	\item{message1}{message1 - the message to be printed - (java expects a RCharacter)}
	\item{message2}{message2 - will be used for toString - (java expects a RCharacter)}
	\if{html}{\out{</div>}}
}


\description{
This has no documentation

Version: 0.01

Generated: 2020-11-03T21:33:22.933
}

\details{
	This has no documentation
}

\examples{\dontrun{
J = testRapi::JavaApi$get();
instance = J$MoreFeatureTest$new(message1, message2)

## -----------------------------------
## Method `MoreFeatureTest$create`
## -----------------------------------
J$MoreFeatureTest$create('Hello,',' World')

## -----------------------------------
## Method `MoreFeatureTest$testLogging`
## -----------------------------------

## -----------------------------------
## Method `MoreFeatureTest$throwCatchable`
## -----------------------------------

## -----------------------------------
## Method `MoreFeatureTest$printMessage`
## -----------------------------------

## -----------------------------------
## Method `MoreFeatureTest$throwRuntime`
## -----------------------------------

## -----------------------------------
## Method `MoreFeatureTest$concat`
## -----------------------------------

}}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$MoreFeatureTest$new(message1, message2)}}
		}
	}
	\subsection{Static methods}{
		\itemize{
			\item \href{#method-create}{\code{J$MoreFeatureTest$create(message1, message2)}}
			\item \href{#method-concat}{\code{J$MoreFeatureTest$concat(message1, message2)}}
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-testLogging}{\code{instance$testLogging()}}
			\item \href{#method-throwCatchable}{\code{instance$throwCatchable()}}
			\item \href{#method-printMessage}{\code{instance$printMessage()}}
			\item \href{#method-throwRuntime}{\code{instance$throwRuntime()}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		the first constructor is used if there are none annotated
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$MoreFeatureTest$new(message1, message2)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{message1 - the message to be printed}{ - (java expects a RCharacter)}
				\item{message2 - will be used for toString}{ - (java expects a RCharacter)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 MoreFeatureTest object: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-create"></a>}}
	
	\subsection{Method \code{create()}}{
		A static object constructor
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$MoreFeatureTest$create(message1, message2)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{message1 - the message to be printed}{ - (java expects a RCharacter)}
				\item{message2 - will be used for toString}{ - (java expects a RCharacter)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			R6 MoreFeatureTest object: A MoreFeatureTest R6 object
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
J$MoreFeatureTest$create('Hello,',' World')
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-testLogging"></a>}}
	
	\subsection{Method \code{testLogging()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$MoreFeatureTest$new(...);
instance$testLogging()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			void: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-throwCatchable"></a>}}
	
	\subsection{Method \code{throwCatchable()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$MoreFeatureTest$new(...);
instance$throwCatchable()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RCharacter: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-printMessage"></a>}}
	
	\subsection{Method \code{printMessage()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$MoreFeatureTest$new(...);
instance$printMessage()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			void: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-throwRuntime"></a>}}
	
	\subsection{Method \code{throwRuntime()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
instance = J$MoreFeatureTest$new(...);
instance$throwRuntime()
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{none}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RCharacter: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-concat"></a>}}
	
	\subsection{Method \code{concat()}}{
		no description
	

		\subsection{Usage}{
			\if{html}{\out{<div class="r">}}
			\preformatted{
J = testRapi::JavaApi$get()
J$MoreFeatureTest$concat(message1, message2)
  			}
			\if{html}{\out{</div>}}
		}
	
		\subsection{Arguments}{
			\if{html}{\out{<div class="arguments">}}
			\itemize{
				\item{message1}{ - (java expects a RCharacter)}
				\item{message2}{ - (java expects a RCharacter)}
			}
			\if{html}{\out{</div>}}
		}

		\subsection{Returns}{
			RCharacter: 
		}

		\subsection{Examples}{
			\if{html}{\out{<div class="r example copy">}}
			\preformatted{
not available
			}
			\if{html}{\out{</div>}}
		}
	}
}