% Generated by r-jsr223-maven-plugin: do not edit by hand
\docType{package}
\name{testRapi-package}
\alias{testRapi}
\alias{testRapi-package}
\title{testRapi: A test library}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}
An optional long description of the package

Version: 0.01

Classes:
\itemize{ 
	\item \code{JavaApi}
	\item \code{TestRApi}
	\item \code{BounceTest}
	\item \code{FactoryTest}
	\item \code{AnotherTestRApi}
}
}

\author{
\itemize{ 
   \item test forename optional surname test@example.com 
   \item r-jsr223-maven-plugin terminological  
   \item given family email@example.com ORCIDID
}
}

\keyword{java api}