% Generated by r6-generator-maven-plugin: do not edit by hand
\docType{package}
\name{testRapi-package}
\alias{testRapi}
\alias{testRapi-package}
\title{testRapi: A test library}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}
This package is a feature test of the R6 Generator Maven Plugin. It is not
							intended to be useful beyond testing, demonstrating and documenting 
							the features of the plugin.

Version: 0.02

Classes:
\itemize{ 
	\item \code{JavaApi}
	\item \code{MoreFeatureTest}
	\item \code{BounceTest}
	\item \code{FeatureTest}
	\item \code{FactoryTest}
	\item \code{Serialiser}
	\item \code{MinimalExample}
}
}

\author{
\itemize{ 
   \item Rob Challen rc538@exeter.ac.uk 
   \item given family email@example.com ORCIDID
   \item terminological   
}
}

\keyword{java api}