% Generated by maven-r-jsr223-plugin: do not edit by hand

\name{TestRApi}

\alias{TestRApi}

\title{}

\usage{\preformatted{
library(testRapi)
J = JavaApi$new();
instance <- J$TestRApi$new(...);
}}

\description{
	A test of the jsr223 templating
}

\details{
	

this is a details comment
}

\examples{
## -----------------------------------
## Method `TestRApi$doHelloWorld`
## -----------------------------------
  An example
  Spans many lines

## -----------------------------------
## Method `TestRApi$fluentSetMessage`
## -----------------------------------

## -----------------------------------
## Method `TestRApi$factoryMethod`
## -----------------------------------

## -----------------------------------
## Method `TestRApi$getMessage`
## -----------------------------------

## -----------------------------------
## Method `TestRApi$doSum`
## -----------------------------------

## -----------------------------------
## Method `TestRApi$objectAsParameter`
## -----------------------------------

## -----------------------------------
## Method `TestRApi$fluentRandomObjectAsParamerer`
## -----------------------------------

## -----------------------------------
## Method `TestRApi$doSomethingWithDataFrame`
## -----------------------------------

## -----------------------------------
## Method `TestRApi$doSomethingWithOtherDataFrame`
## -----------------------------------

## -----------------------------------
## Method `TestRApi$generateRowMajorDataFrame`
## -----------------------------------

## -----------------------------------
## Method `TestRApi$generateColMajorDataFrame`
## -----------------------------------

}

\keyword{java api}

\section{Methods}{
	\subsection{Constructors}{
		\itemize{
			\item \href{#method-new}{\code{J$TestRApi$new()}}
		}
	}
	\subsection{Static methods}{
		\itemize{
		}
	}
	\subsection{Instance methods}{
		\itemize{
			\item \href{#method-doHelloWorld}{\code{instance$doHelloWorld()}}
			\item \href{#method-fluentSetMessage}{\code{instance$fluentSetMessage(message)}}
			\item \href{#method-factoryMethod}{\code{instance$factoryMethod(a, b)}}
			\item \href{#method-getMessage}{\code{instance$getMessage()}}
			\item \href{#method-doSum}{\code{instance$doSum(a, b)}}
			\item \href{#method-objectAsParameter}{\code{instance$objectAsParameter(otherObj)}}
			\item \href{#method-fluentRandomObjectAsParamerer}{\code{instance$fluentRandomObjectAsParamerer(file)}}
			\item \href{#method-doSomethingWithDataFrame}{\code{instance$doSomethingWithDataFrame(colMajorDf)}}
			\item \href{#method-doSomethingWithOtherDataFrame}{\code{instance$doSomethingWithOtherDataFrame(rowMajorDf)}}
			\item \href{#method-generateRowMajorDataFrame}{\code{instance$generateRowMajorDataFrame()}}
			\item \href{#method-generateColMajorDataFrame}{\code{instance$generateColMajorDataFrame()}}
			\item \code{instance$clone()}
			\item \code{instance$print()}
		}
	}

	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-new"></a>}}
	
	\subsection{Method \code{new()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  J$TestRApi$new()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 TestRApi object: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-doHelloWorld"></a>}}
	
	\subsection{Method \code{doHelloWorld()}}{
		Description of a hello world function
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$doHelloWorld()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.String: this java method returns a String
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
  An example
  Spans many lines
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-fluentSetMessage"></a>}}
	
	\subsection{Method \code{fluentSetMessage()}}{
		A fluent method description
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$fluentSetMessage(message)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{message the message is a string}{ - (java expects a java.lang.String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 TestRApi object: nothing as this is a fluent method
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-factoryMethod"></a>}}
	
	\subsection{Method \code{factoryMethod()}}{
		A fluent method description
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$factoryMethod(a, b)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{a}{ - (java expects a java.lang.String)}
				\item{b}{ - (java expects a java.lang.String)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		R6 AnotherTestRApi object: nothing as this is a fluent method
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-getMessage"></a>}}
	
	\subsection{Method \code{getMessage()}}{
		message desciption
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$getMessage()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.String: A message previously set or maybe null
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-doSum"></a>}}
	
	\subsection{Method \code{doSum()}}{
		The doSum function description
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$doSum(a, b)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{a the A parameter}{ - (java expects a java.lang.Double)}
				\item{b the B parameter}{ - (java expects a java.lang.Double)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.Double: A+B of course
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-objectAsParameter"></a>}}
	
	\subsection{Method \code{objectAsParameter()}}{
		
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$objectAsParameter(otherObj)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{otherObj}{ - (java expects a uk.co.terminological.mavenrjsr233plugintest.AnotherTestRApi)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.lang.String: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-fluentRandomObjectAsParamerer"></a>}}
	
	\subsection{Method \code{fluentRandomObjectAsParamerer()}}{
		Should pass and return by reference
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$fluentRandomObjectAsParamerer(file)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{file}{ - (java expects a java.io.File)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.io.File: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-doSomethingWithDataFrame"></a>}}
	
	\subsection{Method \code{doSomethingWithDataFrame()}}{
		Should not pass by reference
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$doSomethingWithDataFrame(colMajorDf)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{colMajorDf}{ - (java expects a java.util.Map<java.lang.String,java.lang.Object[]>)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-doSomethingWithOtherDataFrame"></a>}}
	
	\subsection{Method \code{doSomethingWithOtherDataFrame()}}{
		Should pass by reference unless rowMajor is allowed
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$doSomethingWithOtherDataFrame(rowMajorDf)
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
				\item{rowMajorDf}{ - (java expects a java.util.List<java.util.Map<java.lang.String,java.lang.Object>>)}
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		void: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateRowMajorDataFrame"></a>}}
	
	\subsection{Method \code{generateRowMajorDataFrame()}}{
		Should not pass by reference
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$generateRowMajorDataFrame()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.util.List: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
	\if{html}{\out{<hr>}}
	\if{html}{\out{<a id="method-generateColMajorDataFrame"></a>}}
	
	\subsection{Method \code{generateColMajorDataFrame()}}{
		Should not pass by reference
	}

	\subsection{Usage}{
		\if{html}{\out{<div class="r">}}
		\preformatted{
  library(testRapi)
  J <- JavaApi$new()
  instance <- J$TestRApi$new(...);
  instance$generateColMajorDataFrame()
		}
		\if{html}{\out{</div>}}
	}
	
	\subsection{Arguments}{
		\if{html}{\out{<div class="arguments">}}
		\describe{
			\itemize{
			}
		}
		\if{html}{\out{</div>}}
	}

	\subsection{Returns}{
		java.util.Map: 
	}

	\subsection{Examples}{
		\if{html}{\out{<div class="r example copy">}}
		\preformatted{
		}
		\if{html}{\out{</div>}}
	}
}